Overview

Kinetic.jl is a collection of numerical routines for orthogonal polynomials written in the Julia programming language. Starting from some non-negative weight (aka an absolutely continuous nonnegative measure), PolyChaos allows

to compute the coefficients for the monic three-term recurrence relation,

to evaluate the orthogonal polynomials at arbitrary points,

to compute the quadrature rule,

to compute tensors of scalar products,

to do all of the above in a multivariate setting (aka product measures).

Kinetic.pdf_slope — Function

Calculate slope of particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

pdf_slope(u::Real, Δ::Real)

pdf_slope(prim::AbstractArray{<:Real,1}, sw::AbstractArray{<:Real,1}, inK::Real)

Kinetic.maxwellian — Function

Maxwellian in discrete form

@arg: particle velocity quadrature points

@arg: density, velocity and inverse of temperature

@return: Maxwellian distribution function