Evolution

KitBase.evolve! — Function

evolve!(
    KS::SolverSet,
    ctr::T1,
    face::T2,
    dt;
    mode = Symbol(KS.set.flux)::Symbol,
    bc = :fix::Symbol,
) where {
    T1<:AbstractArray{<:AbstractControlVolume1D,1},
    T2<:AbstractArray{<:AbstractInterface1D,1},
}

evolve!(
    KS::SolverSet,
    ctr::T1,
    a1face::T2,
    a2face::T2,
    dt;
    mode = Symbol(KS.set.flux)::Symbol,
    bc = :fix::Symbol,
) where {
    T1<:AbstractArray{<:AbstractControlVolume2D,2},
    T2<:AbstractArray{<:AbstractInterface2D,2},
}

Evolution of boundary fluxes

The evolution solver calculate the interface numerical fluxes based on two neighbor cells. Different flux functions can be used with the option model.

macroscopic: Godunov, Lax, Roe, HLL, wave-propagation

mesoscopic: upwind, central-upwind, gas-kinetic scheme

The available flux solvers are listed as follows.

KitBase.flux_lax! — Function

flux_lax!(fw::AbstractArray{<:Real,1}, wL::AbstractArray{<:Real,1}, wR::AbstractArray{<:Real,1}, γ::Real, dt::Real, dx::Real)

Lax-Friedrichs flux

P. D. Lax, Weak Solutions of Nonlinear Hyperbolic Equations and Their Numerical Computation, Commun. Pure and Applied Mathematics, 7, 159-193, 1954.

KitBase.flux_hll! — Function

flux_hll!(fw::AbstractArray{<:Real,1}, wL::AbstractArray{<:Real,1}, wR::AbstractArray{<:Real,1}, γ::Real, dt::Real)

HLL flux for the Euler equations

@args: variables at left & right sides of interface

@args: specific heat ratio

KitBase.flux_roe! — Function

flux_roe!(fw::AbstractArray{<:Real,1}, wL::AbstractArray{<:Real,1}, wR::AbstractArray{<:Real,1},
γ::Real, dt::Real, n = [1.0, 0.0]::AbstractArray{<:Real,1})

Roe's flux with entropy fix

P. L. Roe, Approximate Riemann Solvers, Parameter Vectors and Difference Schemes, Journal of Computational Physics, 43, pp. 357-372. (cf. http://cfdbooks.com/cfdcodes.html)

@args primL[1:4] = left state (rhoL, uL, vL, pL)

@args primR[1:4] = right state (rhoR, uR, vR, pR)

@args γ: specific heat ratio

@args n[2]: unit face normal (L -> R)

KitBase.flux_gks — Function

flux_gks(u, μ, dt, a = 0::T, su = 0.0) where {T<:Real}
flux_gks(
    uL::T,
    uR::T,
    μ,
    dt,
    dxL,
    dxR,
    a = 0,
    suL = 0.0,
    suR = 0.0,
) where {T<:Real}

Gas kinetic flux

@args: conservative scalars and their slopes

@args: viscosity

@args: time step and cell size

@return: scalar flux

KitBase.flux_gks! — Function

flux_gks!(
    fw::X,
    wL::Y,
    wR::Y,
    γ::Real,
    inK::Real,
    μᵣ::Real,
    ω::Real,
    dt::Real,
    dxL::Real,
    dxR::Real,
    swL::Y,
    swR::Y,
) where {X<:AbstractArray{<:AbstractFloat,1},Y<:AbstractArray{<:AbstractFloat,1}}

flux_gks!(
    fw::X,
    wL::Y,
    wR::Y,
    inK,
    γ,
    mi,
    ni,
    me,
    ne,
    Kn,
    dt,
    dxL::Real,
    dxR::Real,
    swL::Y,
    swR::Y,
) where {
    X<:AbstractArray{<:AbstractFloat,2},
    Y<:AbstractArray{<:Real,2},
}

flux_gks!(
    fw::X,
    wL::Y,
    wR::Y,
    γ::Real,
    inK::Real,
    μᵣ::Real,
    ω::Real,
    dt::Real,
    dxL::Real,
    dxR::Real,
    dy::Real,
    swL::Y,
    swR::Y,
) where {X<:AbstractArray{<:AbstractFloat,1},Y<:AbstractArray{<:AbstractFloat,1}}

flux_gks!(
    fw::T1,
    fh::T2,
    fb::T2,
    wL::T3,
    wR::T3,
    u::T4,
    inK,
    γ,
    visRef,
    visIdx,
    dt,
    dxL,
    dxR,
    swL::T3,
    swR::T3,
) where {
    T1<:AbstractArray{<:AbstractFloat,1},
    T2<:AbstractArray{<:AbstractFloat,1},
    T3<:AbstractArray{<:Real,1},
    T4<:AbstractArray{<:AbstractFloat,1},
}

flux_gks!(
    fw::X,
    wL::Y,
    wR::Y,
    inK,
    γ,
    mi,
    ni,
    me,
    ne,
    Kn,
    dt,
    dxL,
    dxR,
    len,
    swL::Y,
    swR::Y,
) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,2}}

Gas kinetic Navier-Stokes flux

@args: conservative variables and their left/right slopes

@args: molecular and thermodynamic parameters

@args: time step and cell size

KitBase.flux_kfvs! — Function

DOM: `flux_kfvs!(ff, fL, fR, u, dt, sfL, sfR)`
1D1F1V: `flux_kfvs!(fw, ff, fL, fR, u, ω, dt, sfL, sfR)`
1D1F3V: `flux_kfvs!(fw, ff, fL, fR, u, v, w, ω, dt, sfL, sfR)`
1D2F1V: `flux_kfvs!(fw, fh, fb, hL, bL, hR, bR, u, ω, dt, shL, sbL, shR, sbR)`
1D4F1V: `flux_kfvs!(fw, fh0, fh1, fh2, fh3, h0L, h1L, h2L, h3L, h0R, h1R, h2R, h3R, u, ω, dt, sh0L, sh1L, sh2L, sh3L, sh0R, sh1R, sh2R, sh3R)`
2D1F2V: `flux_kfvs!(fw, ff, fL, fR, u, v, ω, dt, len, sfL, sfR)`
2D2F2V: `flux_kfvs!(fw, fh, fb, hL, bL, hR, bR, u, v, ω, dt, len, shL, sbL, shR, sbR)`

Kinetic flux vector splitting (KFVS) flux

@args: particle distribution functions and their left/right slopes

@args: particle velocity quadrature points and weights

@args: time step and cell size

KitBase.flux_kcu! — Function

1D1F1V: flux_kcu!(fw, ff, wL, fL, wR, fR, u, ω, inK, γ, visRef, visIdx, Pr, dt)
1D2F1V: flux_kcu!(fw, fh, fb, wL, hL, bL, wR, hR, bR, u, ω, inK, γ, visRef, visIdx, Pr, dt)
1D4F1V: flux_kcu!(fw, fh0, fh1, fh2, fh3, wL, h0L, h1L, h2L, h3L, wR, h0R, h1R, h2R, h3R, u, ω, inK, γ, visRef, visIdx, Pr, dt)
2D1F2V: flux_kcu!(fw, ff, wL, fL, wR, fR, u, v, ω, inK, γ, visRef, visIdx, Pr, dt, len)
2D2F2V: flux_kcu!(fw, fh, fb, wL, hL, bL, wR, hR, bR, u, v, ω, inK, γ, visRef, visIdx, Pr, dt, len)
2D3F2V: flux_kcu!(fw, fh0, fh1, fh2, wL, h0L, h1L, h2L, wR, h0R, h1R, h2R, u, v, ω, inK, γ, visRef, visIdx, Pr, dt, len)

Kinetic central-upwind (KCU) method

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step and cell size

KitBase.flux_ugks! — Function

flux_ugks!(
    fw::T1,
    fh::T2,
    fb::T2,
    wL::T3,
    hL::T4,
    bL::T4,
    wR::T3,
    hR::T4,
    bR::T4,
    u::T5,
    ω::T5,
    inK,
    γ,
    visRef,
    visIdx,
    pr,
    dt,
    dxL,
    dxR,
    shL = zeros(eltype(hL), axes(hL))::T4,
    sbL = zeros(eltype(bL), axes(bL))::T4,
    shR = zeros(eltype(hR), axes(hR))::T4,
    sbR = zeros(eltype(bR), axes(bR))::T4,
) where {
    T1<:AbstractArray{<:AbstractFloat,1},
    T2<:AbstractArray{<:AbstractFloat,1},
    T3<:AbstractArray{<:Real,1},
    T4<:AbstractArray{<:AbstractFloat,1},
    T5<:AbstractArray{<:AbstractFloat,1},
}

flux_ugks!(
    fw::T1,
    fh::T2,
    fb::T2,
    wL::T3,
    hL::T4,
    bL::T4,
    wR::T3,
    hR::T4,
    bR::T4,
    u::T5,
    v::T5,
    ω::T5,
    inK,
    γ,
    visRef,
    visIdx,
    pr,
    dt,
    dxL,
    dxR,
    len,
    shL = zeros(eltype(hL), axes(hL))::T4,
    sbL = zeros(eltype(bL), axes(bL))::T4,
    shR = zeros(eltype(hR), axes(hR))::T4,
    sbR = zeros(eltype(bR), axes(bR))::T4,
) where {
    T1<:AbstractArray{<:AbstractFloat,1},
    T2<:AbstractArray{<:AbstractFloat,2},
    T3<:AbstractArray{<:Real,1},
    T4<:AbstractArray{<:AbstractFloat,2},
    T5<:AbstractArray{<:AbstractFloat,2},
}

flux_ugks!(
    fw::T1,
    fh0::T2,
    fh1::T2,
    fh2::T2,
    wL::T3,
    h0L::T4,
    h1L::T4,
    h2L::T4,
    wR::T3,
    h0R::T4,
    h1R::T4,
    h2R::T4,
    u::T5,
    v::T5,
    ω::T5,
    inK,
    γ,
    mi,
    ni,
    me,
    ne,
    Kn,
    dt,
    dxL,
    dxR,
    len,
    sh0L = zeros(eltype(h0L), axes(h0L))::T4,
    sh1L = zeros(eltype(h1L), axes(h1L))::T4,
    sh2L = zeros(eltype(h2L), axes(h2L))::T4,
    sh0R = zeros(eltype(h0R), axes(h0R))::T4,
    sh1R = zeros(eltype(h1R), axes(h1R))::T4,
    sh2R = zeros(eltype(h2R), axes(h2R))::T4,
) where {
    T1<:AbstractArray{<:AbstractFloat,2},
    T2<:AbstractArray{<:AbstractFloat,3},
    T3<:AbstractArray{<:Real,2},
    T4<:AbstractArray{<:AbstractFloat,3},
    T5<:AbstractArray{<:AbstractFloat,3},
}

Unified gas kinetic scheme (UGKS)

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step

KitBase.flux_boundary_maxwell! — Function

flux_boundary_maxwell!(
    fw::T1,
    fh::T2,
    fb::T2,
    bc::T3,
    h::T4,
    b::T4,
    u::T5,
    ω::T5,
    inK,
    dt,
    rot = 1,
) where {
    T1<:AbstractArray{<:AbstractFloat,1},
    T2<:AbstractArray{<:AbstractFloat,1},
    T3<:Array{<:Real,1},
    T4<:AbstractArray{<:AbstractFloat,1},
    T5<:AbstractArray{<:AbstractFloat,1},
}

flux_boundary_maxwell!(
    fw::T1,
    fh::T2,
    fb::T2,
    bc::T3,
    h::T4,
    b::T4,
    u::T5,
    v::T5,
    ω::T5,
    inK,
    dt,
    len,
    rot = 1,
) where {
    T1<:AbstractArray{<:AbstractFloat,1},
    T2<:AbstractArray{<:AbstractFloat,2},
    T3<:Array{<:Real,1},
    T4<:AbstractArray{<:AbstractFloat,2},
    T5<:AbstractArray{<:AbstractFloat,2},
}

Maxwell's diffusive boundary flux

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step

KitBase.flux_boundary_specular! — Function

flux_boundary_specular!(
    fw::T1,
    ff::T2,
    f::T3,
    u::T4,
    ω::T4,
    dt,
) where {
    T1<:AbstractArray{<:Real,1},
    T2<:AbstractArray{<:AbstractFloat,1},
    T3<:AbstractArray{<:AbstractFloat,1},
    T4<:AbstractArray{<:AbstractFloat,1},
}

flux_boundary_specular!(
    fw::T1,
    fh::T2,
    fb::T2,
    h::T3,
    b::T3,
    u::T4,
    ω::T4,
    dt,
) where {
    T1<:AbstractArray{<:Real,1},
    T2<:AbstractArray{<:AbstractFloat,1},
    T3<:AbstractArray{<:AbstractFloat,1},
    T4<:AbstractArray{<:AbstractFloat,1},
}

Specular reflection boundary flux

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step

KitBase.flux_em! — Function

flux_em!(femL, femR, ELL, BLL, EL, BL, ER, BR, ERR, BRR, ϕL, ϕR, ψL, ψR, dxL, dxR, Ap, An, D, sol, χ, ν, dt)

Wave propagation method for Maxwell's equations

@args: {E, B, ϕ, ψ} in left-left, left, right, and right-right cells

@args: eigenmatrix (A -> A+ & A-), eigenvalue (D)

@args: full size of left & right cells

@args: speed of light (sol)

@args: auxiliary parameters (χₑ, νᵦ)

KitBase.flux_emx! — Function

flux_emx!(femL, femR, femLU, femLD, femRU, femRD,
ELL, BLL, EL, BL, ER, BR, ERR, BRR, ϕL, ϕR, ψL, ψR,
dxL, dxR, A1p, A1n, A2p, A2n, D, sol, χ, ν, dt)

Wave propagation method for 2D Maxwell's equations

@args: {E, B, ϕ, ψ} in left-left, left, right, and right-right cells

@args: eigenmatrix (A -> A+ & A-), eigenvalue (D)

@args: full size of left & right cells

@args: speed of light (sol)

@args: auxiliary parameters (χₑ, νᵦ)

KitBase.flux_emy! — Function

flux_emx!(femL, femR, femLU, femLD, femRU, femRD,
ELL, BLL, EL, BL, ER, BR, ERR, BRR, ϕL, ϕR, ψL, ψR,
dxL, dxR, A1p, A1n, A2p, A2n, D, sol, χ, ν, dt)

Wave propagation method for 2D Maxwell's equations

@args: {E, B, ϕ, ψ} in left-left, left, right, and right-right cells

@args: eigenmatrix (A -> A+ & A-), eigenvalue (D)

@args: full size of left & right cells

@args: speed of light (sol)

@args: auxiliary parameters (χₑ, νᵦ)