Timestep

KitBase.timestep — Function

timestep(KS, ctr, simTime)

Calculate timestep

@return: Δt

The timestep solver returns the time interval used for the upcoming solution loop based on the current variables.