Data Structure

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume1D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume1D1F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume1D2F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume1D3F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume1D4F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume2D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume2D1F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume2D2F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.ControlVolume2D3F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface1D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface1D1F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface1D2F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface1D3F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface1D4F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface2D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface2D1F. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic.KitBase.Interface2D2F. Check Documenter's build log for details.

KitBase.aap_hs_collision_time — Method

Calculate mixture collision time from AAP model

aap_hs_collision_time( prim::AbstractArray{<:Real,2}, mi::Real, ni::Real, me::Real, ne::Real, kn::Real, )

KitBase.aap_hs_diffeq! — Method

Source term of AAP model in DifferentialEquations.jl

aap_hs_diffeq!(du, u, p, t)

KitBase.aap_hs_prim — Method

Calculate mixture primitive variables from AAP model

tested √

aap_hs_prim( prim::AbstractArray{<:Real,2}, tau::AbstractArray{<:Real,1}, mi::Real, ni::Real, me::Real, ne::Real, kn::Real, )

KitBase.advection_flux — Method

Theoretical flux of linear advection equation

advection_flux(u, a)

KitBase.angle — Method

Args order (B,A,C) isn't a mistake

KitBase.boltzmann_fft! — Method

Calculate collision operator with FFT-based fast spectral method

KitBase.boltzmann_fft — Method

Calculate collision operator with FFT-based fast spectral method

KitBase.burgers_flux — Method

Theoretical flux of Burgers' equation

burgers_flux(u)

KitBase.central_diff! — Method

Central difference

central_diff!(dy::AbstractArray{<:Any,1}, y::AbstractArray{<:Any,1}, x::AbstractArray{<:Any,1})

central_diff!(dy::AbstractArray{<:Any,1}, y::AbstractArray{<:Any,1}, dx::Any)

KitBase.central_diff — Method

Central difference

central_diff(y::AbstractArray{<:Any,1}, x::AbstractArray{<:Any,1})

central_diff(y::AbstractArray{<:Any,1}, dx::Any)

KitBase.conserve_prim — Method

Transform conservative -> primitive variables

scalar: pseudo primitive vector for scalar conservation laws

conserve_prim(u)

conserve_prim(u, a)

vector: primitive vector for Euler, Navier-Stokes and extended equations

conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,1}}

conserve_prim(ρ, M, E, γ)

conserve_prim(ρ, MX, MY, E, γ)

KitBase.discrete_moments — Method

Discrete moments of particle distribution

discrete_moments(f, ω): direct quadrature

discrete_moments(f, u, ω, n): velocity moments

KitBase.em_coefficients — Method

Calculate electromagnetic coeffcients in hyperbolic Maxwell's equations

em_coefficients(
    prim::X,
    E::Y,
    B::Z,
    mr,
    lD,
    rL,
    dt,
) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:Real,1}}

KitBase.euler_flux — Method

Theoretical fluxes of Euler Equations

euler_flux(w::A, γ; frame = :cartesian::Symbol) where {A<:AbstractArray{<:Real,1}}

@return: flux tuple

KitBase.euler_jacobi — Method

Flux Jacobian of Euler Equations

euler_jacobi(w::T, γ) where {T<:AbstractArray{<:Real,1}}

@return: Jacobian matrix A

KitBase.evolve! — Method

Evolution

KitBase.extract_last — Method

Extract subarray except the last column

extract_last(a::AbstractArray, idx::Int; mode=:view::Symbol)

KitBase.flux_boundary_maxwell! — Method

Maxwell's diffusive boundary flux

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step

KitBase.flux_em! — Method

Wave propagation method for Maxwell's equations

flux_em!(femL, femR, ELL, BLL, EL, BL, ER, BR, ERR, BRR, ϕL, ϕR, ψL, ψR, dxL, dxR, Ap, An, D, sol, χ, ν, dt)

@args: {E, B, ϕ, ψ} in left-left, left, right, and right-right cells

@args: eigenmatrix (A -> A+ & A-), eigenvalue (D)

@args: full size of left & right cells

@args: speed of light (sol)

@args: auxiliary parameters (χₑ, νᵦ)

KitBase.flux_emx! — Method

Wave propagation method for 2D Maxwell's equations

flux_emx!(femL, femR, femLU, femLD, femRU, femRD,
ELL, BLL, EL, BL, ER, BR, ERR, BRR, ϕL, ϕR, ψL, ψR,
dxL, dxR, A1p, A1n, A2p, A2n, D, sol, χ, ν, dt)

@args: {E, B, ϕ, ψ} in left-left, left, right, and right-right cells

@args: eigenmatrix (A -> A+ & A-), eigenvalue (D)

@args: full size of left & right cells

@args: speed of light (sol)

@args: auxiliary parameters (χₑ, νᵦ)

KitBase.flux_emy! — Method

Wave propagation method for 2D Maxwell's equations

flux_emx!(femL, femR, femLU, femLD, femRU, femRD,
ELL, BLL, EL, BL, ER, BR, ERR, BRR, ϕL, ϕR, ψL, ψR,
dxL, dxR, A1p, A1n, A2p, A2n, D, sol, χ, ν, dt)

@args: {E, B, ϕ, ψ} in left-left, left, right, and right-right cells

@args: eigenmatrix (A -> A+ & A-), eigenvalue (D)

@args: full size of left & right cells

@args: speed of light (sol)

@args: auxiliary parameters (χₑ, νᵦ)

KitBase.flux_gks! — Method

Gas kinetic Navier-Stokes flux

1D: flux_gks!(fw, wL, wR, γ, K, μᵣ, ω, dt, dx, swL, swR)
2D: flux_gks!(fw, wL, wR, γ, K, μᵣ, ω, dt, dx, dy, swL, swR)

@args: conservative variables and their left/right slopes

@args: molecular and thermodynamic parameters

@args: time step and cell size

KitBase.flux_gks — Method

Gas kinetic Navier-Stokes flux

flux_gks(u::Real, μ::Real, dt::Real, su = 0.0::Real, a = 0::Real)
flux_gks(uL::Real, uR::Real, μ::Real, dt::Real, dxL::Real,
    dxR::Real, suL = 0.0::Real, suR = 0.0::Real, a = 0::Real)

@args: conservative scalars and their slopes

@args: viscosity

@args: time step and cell size

@return: scalar flux

KitBase.flux_hll! — Method

HLL flux for the Euler equations

flux_hll!(fw::AbstractArray{<:Real,1}, wL::AbstractArray{<:Real,1}, wR::AbstractArray{<:Real,1}, γ::Real, dt::Real)

@args: variables at left & right sides of interface

@args: specific heat ratio

KitBase.flux_kcu! — Method

Kinetic central-upwind (KCU) method

1D1F1V: flux_kcu!(fw, ff, wL, fL, wR, fR, u, ω, inK, γ, visRef, visIdx, Pr, dt)
1D2F1V: flux_kcu!(fw, fh, fb, wL, hL, bL, wR, hR, bR, u, ω, inK, γ, visRef, visIdx, Pr, dt)
1D4F1V: flux_kcu!(fw, fh0, fh1, fh2, fh3, wL, h0L, h1L, h2L, h3L, wR, h0R, h1R, h2R, h3R, u, ω, inK, γ, visRef, visIdx, Pr, dt)
2D1F2V: flux_kcu!(fw, ff, wL, fL, wR, fR, u, v, ω, inK, γ, visRef, visIdx, Pr, dt, len)
2D2F2V: flux_kcu!(fw, fh, fb, wL, hL, bL, wR, hR, bR, u, v, ω, inK, γ, visRef, visIdx, Pr, dt, len)
2D3F2V: flux_kcu!(fw, fh0, fh1, fh2, wL, h0L, h1L, h2L, wR, h0R, h1R, h2R, u, v, ω, inK, γ, visRef, visIdx, Pr, dt, len)

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step and cell size

KitBase.flux_kfvs! — Method

Kinetic flux vector splitting (KFVS) flux

DOM: `flux_kfvs!(ff, fL, fR, u, dt, sfL, sfR)`
1D1F1V: `flux_kfvs!(fw, ff, fL, fR, u, ω, dt, sfL, sfR)`
1D1F3V: `flux_kfvs!(fw, ff, fL, fR, u, v, w, ω, dt, sfL, sfR)`
1D2F1V: `flux_kfvs!(fw, fh, fb, hL, bL, hR, bR, u, ω, dt, shL, sbL, shR, sbR)`
1D4F1V: `flux_kfvs!(fw, fh0, fh1, fh2, fh3, h0L, h1L, h2L, h3L, h0R, h1R, h2R, h3R, u, ω, dt, sh0L, sh1L, sh2L, sh3L, sh0R, sh1R, sh2R, sh3R)`
2D1F2V: `flux_kfvs!(fw, ff, fL, fR, u, v, ω, dt, len, sfL, sfR)`
2D2F2V: `flux_kfvs!(fw, fh, fb, hL, bL, hR, bR, u, v, ω, dt, len, shL, sbL, shR, sbR)`

@args: particle distribution functions and their left/right slopes

@args: particle velocity quadrature points and weights

@args: time step and cell size

KitBase.flux_lax! — Method

Lax-Friedrichs flux

flux_lax!(fw::AbstractArray{<:Real,1}, wL::AbstractArray{<:Real,1}, wR::AbstractArray{<:Real,1}, γ::Real, dt::Real, dx::Real)

P. D. Lax, Weak Solutions of Nonlinear Hyperbolic Equations and Their Numerical Computation, Commun. Pure and Applied Mathematics, 7, 159-193, 1954.

KitBase.flux_roe! — Method

Roe's flux with entropy fix

flux_roe!(fw::AbstractArray{<:Real,1}, wL::AbstractArray{<:Real,1}, wR::AbstractArray{<:Real,1},
γ::Real, dt::Real, n = [1.0, 0.0]::AbstractArray{<:Real,1})

P. L. Roe, Approximate Riemann Solvers, Parameter Vectors and Difference Schemes, Journal of Computational Physics, 43, pp. 357-372. (cf. http://cfdbooks.com/cfdcodes.html)

@args primL[1:4] = left state (rhoL, uL, vL, pL)

@args primR[1:4] = right state (rhoR, uR, vR, pR)

@args γ: specific heat ratio

@args n[2]: unit face normal (L -> R)

KitBase.flux_ugks! — Method

Unified gas kinetic scheme (UGKS)

@args: particle distribution functions and their slopes at left/right sides of interface

@args: particle velocity quadrature points and weights

@args: time step

KitBase.fortsign — Method

Fortran sign function

fortsign(x::Real, y::Real)

KitBase.full_distribution — Method

Recover full distribution function from reduced ones

@arg h & b : reduced particle distribution function with 1D velocity space

@arg u : quadrature nodes in 1D velocity space

@arg weights : quadrature weights in 1D velocity space

@arg v & w : quadrature nodes in the rest velocity space (with 3D setting)

@return f : particle distribution function with 3D velocity space

KitBase.gauss_moments — Method

Calculate moments of Gaussian distribution G = (λ / π)^(D / 2) * exp[-λ(c^2 + ξ^2)]

internality: gauss_moments(prim::T) where {T<:AbstractArray{<:Real,1}}

no internality: gauss_moments(prim::T, inK) where {T<:AbstractArray{<:Real,1}}

KitBase.global_frame — Method

Transform local flow variables to global frame

2D: global_frame(w::AbstractArray{<:Real,1}, cosa::Real, sina::Real)

3D: global_frame(w::AbstractArray{<:Real,1}, dirccos::AbstractArray{<:Real,2})

KitBase.heat_capacity_ratio — Method

Calculate heat capacity ratio

KitBase.heat_flux — Method

Calculate heat flux from particle distribution function

KitBase.heaviside — Method

Heaviside step function

heaviside(x::Real)

KitBase.hs_boltz_kn — Method

Calculate effective Knudsen number for fast spectral method

hard sphere (HS) model

KitBase.ib_briowu — Method

Initialize Brio-Wu MHD shock

ib_briowu(gam, uspace::T, mi, me) where {T<:AbstractArray{<:AbstractFloat,2}}

KitBase.ib_cavity — Method

Initialize lid-driven cavity

2d1f2v: ib_cavity( gam, Um, Vm, Tm, u::T, v::T) where {T<:AbstractArray{<:AbstractFloat,2}}

2d2f2v: ib_cavity( gam, Um, Vm, Tm, u::T, v::T, K) where {T<:AbstractArray{<:AbstractFloat,2}}

KitBase.ib_rh — Method

Initialize Rankine-Hugoniot relation

1d1f1v: ib_rh(MaL, gam, u::T) where {T<:AbstractArray{<:AbstractFloat,1}}

1d2f1v: ib_rh(MaL, gam, u::T, K) where {T<:AbstractArray{<:AbstractFloat,1}}

1d1f3v: ib_rh(MaL, gam, u::T, v::T, w::T) where {T<:AbstractArray{<:AbstractFloat,3}}

KitBase.ib_sod — Method

Initialize Sod shock tube

1d1f1v: ib_sod(γ, u::T) where {T<:AbstractArray{<:AbstractFloat,1}}

1d1f3v: ib_sod(γ, u::T, v::T, w::T) where {T<:AbstractArray{<:AbstractFloat,3}}

1d2f1v: ib_sod(γ, u::T, K) where {T<:AbstractArray{<:AbstractFloat,1}}

KitBase.init_fvm — Method

Initialize finite volume method

KitBase.initialize — Method

Initialize solver from input file

KitBase.kernel_mode — Method

Calculate collision kernel for fast spectral method

KitBase.legendre_quadrature — Method

Gauss-Legendre quadrature

legendre_quadrature(n::Int)

@arg n : quadrature order (MUST be even)

@return points : quadrature points in 3D coordinate

@return weights : quadrature weights

KitBase.lgwt — Method

Gauss Legendre integral for fast spectral method

lgwt(N::Int, a::Real, b::Real)

@args: number of quadrature points N, integral range [a, b]

@args: quadrature points x & weights w

KitBase.linspace — Method

Python linspace function

linspace(start::Real, stop::Real, n::Int)

KitBase.local_frame — Method

Transform global flow variables to local frame

2D: local_frame(w::AbstractArray{<:Real,1}, cosa::Real, sina::Real)

3D: local_frame(w::AbstractArray{<:Real,1}, dirccos::AbstractArray{<:Real,2})

KitBase.mat_split — Method

Split matrix into row vectors

mat_split(m::AbstractArray)

KitBase.maxwellian — Method

Maxwellian in discrete form

@args: particle velocity quadrature points

@args: density, velocity and inverse of temperature

@return: Maxwellian distribution function

KitBase.mesh_area_2D — Method

Compute areas of 2D elements

mesh_area_2D(nodes::AbstractArray{<:AbstractFloat,2}, cells::AbstractArray{<:Int,2})

KitBase.mesh_center_2D — Method

Compute central points of 2D elements

mesh_center_2D(nodes::AbstractArray{<:AbstractFloat,2}, cells::AbstractArray{<:Int,2})

KitBase.mesh_connectivity_2D — Method

Compute connectivity of 2D unstructured mesh

mesh_connectivity_2D(cells::AbstractArray{<:Int,2})

KitBase.meshgrid — Method

Equivalent structured mesh generator as matlab

2D: meshgrid(x::AbstractArray{<:Real,1}, y::AbstractArray{<:Real,1})

3D: meshgrid(x::AbstractArray{<:Real,1}, y::AbstractArray{<:Real,1}, z::AbstractArray{<:Real,1})

KitBase.minmod — Method

minmod limiter

minmod(sL::Real, sR::Real)

KitBase.mixture_conserve_prim — Method

Transform multi-component conservative -> primitive variables

mixture_conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,2}}

KitBase.mixture_gauss_moments — Method

Calculate moments of Gaussian distribution in multi-component gas

mixture_gauss_moments(prim::T, inK) where {T<:AbstractArray{<:Real,2}}

KitBase.mixture_maxwellian — Method

Multi-component Maxwellian in discrete form

mixture_maxwellian(u::X, prim::Y) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,2}}

mixture_maxwellian(
    u::X,
    v::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,3},Y<:AbstractArray{<:Real,2}}

mixture_maxwellian(
    u::X,
    v::X,
    w::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,4},Y<:AbstractArray{<:Real,2}}

KitBase.mixture_moments_conserve — Method

Calculate conservative moments of particle distribution in multi-component gas

mixture_moments_conserve(
    Mu::T,
    Mxi::T,
    alpha::I,
    delta::I,
) where {T<:OffsetArray{<:AbstractFloat,2},I<:Int}

function mixture_moments_conserve(
    Mu::T,
    Mv::T,
    Mw::T,
    alpha::I,
    beta::I,
    delta::I,
) where {T<:OffsetArray{<:AbstractFloat,2},I<:Int}

KitBase.mixture_moments_conserve_slope — Method

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.mixture_pdf_slope — Method

Calculate slope of multi-component particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

mixture_pdf_slope(prim::X, sw::Y, inK) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,2}}

KitBase.mixture_prim_conserve — Method

Transform multi-component primitive -> conservative variables

mixture_prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,2}}

KitBase.moments_conserve — Method

Calculate conservative moments of particle distribution

moments_conserve(Mu::OffsetArray{<:AbstractFloat,1}, alpha::Int)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mxi::OffsetArray{<:Real,1}, alpha::Int, delta::Int)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mv::OffsetArray{<:Real,1}, Mw::OffsetArray{<:Real,1}, alpha::Int, beta::Int, delta::Int)

KitBase.moments_conserve_slope — Method

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.newton_cotes — Method

Newton-Cotes rule

newton_cotes(idx::T, num::T) where {T<:Int}

KitBase.octa_quadrature — Method

Octaeder quadrature

octa_quadrature(n::Int, slerpflag = true::Bool)

@arg n : quadrature order

@arg slerpflag : flag of spherical linear interpolation

@return points

@return triangulation

KitBase.pdf_slope — Method

Calculate slope of particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

pdf_slope(u, Δ)

pdf_slope(prim::A, sw::B, inK) where {A<:AbstractArray{<:Real,1},B<:AbstractArray{<:Real,1}}

KitBase.plot_line — Method

Plot 1D profile

plot_line(KS, ctr; backend)

KitBase.prim_conserve — Method

Transform primitive -> conservative variables

prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,1}}

prim_conserve(ρ, U, λ, γ)

prim_conserve(ρ, U, V, λ, γ)

prim_conserve(ρ, U, V, W, λ, γ)

KitBase.quadrature_weights — Method

Create quadrature weights from points and triangulation

create_weights(n::Int, xyz::AbstractArray{<:Real,2}, triangles::AbstractArray{Int,2})

@arg xyz : quadrature points

@arg triangles : triangulation

@return weights : quadrature weights

KitBase.read_dict — Method

Read text into dictionary

read_dict(filename::String, allowed)
read_dict(filename::String)

@args filename: configuration text file

@args allowed: keywords

@return vars: dictionary with values of variables

KitBase.read_mesh — Method

Read mesh file

read_mesh(file)

@return nodes : are saved with 3D coordinates (z=0 for 2D case)

@return cells : node ids inside cells

KitBase.reconstruct! — Method

Reconstructor

1D solver: reconstruct!(KS::SolverSet, ctr::AbstractArray{<:AbstractControlVolume1D,1})

2D solver: reconstruct!(KS::SolverSet, ctr::AbstractArray{ControlVolume2D2F,2})

KitBase.reconstruct2! — Method

Two-cell reconstruction

KitBase.reconstruct2 — Method

Two-cell reconstruction

KitBase.reconstruct3! — Method

Three-cell reconstruction

KitBase.reconstruct3 — Method

Three-cell reconstruction

KitBase.reduce_distribution — Method

Reduced distribution function

@arg : particle distribution function with full velocity space

@arg : quadrature weights with reduced velocity setting (v & w by default)

KitBase.ref_vhs_vis — Method

Calculate reference viscosity

variable hard sphere (VHS) model

KitBase.set_ib — Method

Generate AbstractIB

KitBase.shakhov — Method

Shakhov non-equilibrium part

@arg: particle velocity quadrature points

@arg: discrete Maxwellian

@arg: primitive variables, Prandtl number, heat flux, inner degree of freedom

KitBase.shift_pdf! — Method

Shift distribution function by external force

shift_pdf!(
    f::T,
    a,
    du,
    dt,
) where {T<:AbstractArray{<:AbstractFloat,1}}

shift_pdf!(
    f::X,
    a::Y,
    du::Z,
    dt,
) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:AbstractFloat,1}}

KitBase.slerp — Method

Spherical linear interpolation

slerp(pt1::T, pt2::T, n::I) where {T<:AbstractArray{<:Real,1},I<:Int}

KitBase.solve! — Method

Solution algorithm

1D solver: solve!(KS::SolverSet, ctr::AbstractArray{<:AbstractControlVolume1D,1}, face::Array{<:AbstractInterface1D,1}, simTime::Float64)

@return: ending time

KitBase.sound_speed — Method

Calculate speed of sound

KitBase.step! — Method

Update flow variables with finite volume formulation

wrapper: step!(ks, faceL, cell, faceR, dt, res, avg, collision=:bgk, isMHD=:true)

1d0f: step!(fwL, w, prim, fwR, γ, dx, RES, AVG)

1d1f1v: step!(fwL, ffL, w, prim, f, fwR, ffR, u, weights, γ, μᵣ, ω, Pr, dx, dt, RES, AVG, collision=:bgk)

1d1f3v: step!(fwL, ffL, w, prim, f, fwR, ffR, uVelo, vVelo, wVelo, weights, γ, μᵣ, ω, Pr, dx, dt, RES, AVG, collision=:bgk)

1d2f1v: step!(fwL, fhL, fbL, w, prim, h, b, fwR, fhR, fbR, u, weights, K, γ, μᵣ, ω, Pr, dx, dt, RES, AVG, collision=:bgk)

1d2f1v2s: step!(fwL, fhL, fbL, w, prim, h, b, fwR, fhR, fbR, u, weights, K, γ, mi, ni, me, ne, Kn, Pr, dx, dt, RES, AVG, collision=:bgk)

KitBase.stress — Method

Calculate stress tensor from particle distribution function

KitBase.superbee — Method

superbee limiter

superbee(sL::Real, sR::Real)

KitBase.timestep — Method

Timestep calculator

1D solver: timestep(KS::SolverSet, ctr::AbstractArray{<:AbstractControlVolume1D,1}, simTime::Real)

@return: Δt

KitBase.uniform_mesh — Method

Generate uniform mesh

uniform_mesh(x0::Real, xnum::Int, dx::Real)

KitBase.unique — Method

Cleaner for all duplicate (non unique) entries of quadrature points and triangles

unique(Points::X, Triangles::Y) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Int,2}}

@arg Points : quadrature points

@arg Triangles : triangulation

@return xyz & triangulation : new quadrature points and triangulation

KitBase.unstruct_diff — Method

Finite difference for pseudo-unstructured mesh

unstruct_diff(
    u::AbstractArray{<:Any,1},
    x::AbstractArray{<:Any,1},
    nx::Int;
    mode = :central::Symbol,
)

unstruct_diff(u::Function, x::AbstractArray{<:Any,2}, nx::Int, dim::Int; mode = :central::Symbol)

KitBase.update! — Method

Update flow variables

KitBase.upwind_diff! — Method

Upwind difference

upwind_diff!(
    dy::AbstractArray{<:Any,1},
    y::AbstractArray{<:Any,1},
    x::AbstractArray{<:Any,1};
    stream = :right::Symbol,
)

upwind_diff!(
    dy::AbstractArray{<:Any,1},
    y::AbstractArray{<:Any,1},
    dx::Any;
    stream = :right::Symbol,
)

KitBase.upwind_diff — Method

Upwind difference

upwind_diff(
    y::AbstractArray{<:Any,1},
    x::AbstractArray{<:Any,1};
    stream = :right::Symbol,
)

upwind_diff(y::AbstractArray{<:Any,1}, dx::Any; stream = :right::Symbol)

KitBase.vanalbaba — Method

van Albaba limiter

vanalbaba(sL::Real, sR::Real)

KitBase.vanleer — Method

van Leer limiter

vanleer(sL::Real, sR::Real)

KitBase.vhs_collision_time — Method

Calculate collision time

variable hard sphere (VHS) model

KitBase.weno5 — Method

5th-order WENO-JS interpolation

weno5(wL2::T, wL1::T, wN::T, wR1::T, wR2::T) where {T}

KitBase.write_jld — Method

Write data into JLD2

write_jld(KS, ctr, t)

KitBase.ControlVolume1D — Type

1D control volume with no distribution function

ControlVolume1D(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray)

@vars: x, dx, w, prim, sw

KitBase.ControlVolume1D1F — Type

1D control volume with 1 distribution function

ControlVolume1D1F(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray, F::AbstractArray)

@vars: x, dx, w, prim, sw, f, sf

KitBase.ControlVolume1D2F — Type

1D control volume with 2 distribution functions

ControlVolume1D2F(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray, H::AbstractArray, B::AbstractArray)

@vars: x, dx, w, prim, sw, h, b, sh, sb

KitBase.ControlVolume1D3F — Type

1D control volume with 3 distribution functions

ControlVolume1D3F(X, DX, W, PRIM, H0, H1, H2, E0, B0, L)

@vars: x, dx, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

KitBase.ControlVolume1D4F — Type

1D control volume with 4 distribution functions

ControlVolume1D4F(X, DX, W, PRIM, H0, H1, H2, H3, E0, B0, L)

@vars: x, dx, w, prim, sw, h0, h1, h2, h3, sh0, sh1, sh2, sh3, E, B, ϕ, ψ, lorenz

KitBase.ControlVolume2D — Type

2D control volume with no distribution function

ControlVolume2D(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw

KitBase.ControlVolume2D1F — Type

2D control volume with 1 distribution function

ControlVolume2D1F(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray, F::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw, f, sf

KitBase.ControlVolume2D2F — Type

2D control volume with 2 distribution functions

ControlVolume2D2F(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray, H::AbstractArray, B::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw, h, b, sh, sb

KitBase.ControlVolume2D3F — Type

2D control volume with 3 distribution functions

ControlVolume2D3F(X, DX, Y, DY, W, PRIM, H0, H1, H2, E0, B0, L)

@vars: x, y, dx, dy, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

KitBase.Gas — Type

Gas property

@consts: Kn, Ma, Pr, K, γ, ω, αᵣ, ωᵣ, μᵣ

KitBase.IB — Type

Initial & boundary condition with no distribution function

@consts: wL, primL, bcL, wR, primR, bcR, bcU, bcD

KitBase.IB1F — Type

Initial & boundary condition with 1 distribution function

@consts: wL, primL, fL, bcL, wR, primR, fR, bcR, bcU, bcD

KitBase.IB2F — Type

Initial & boundary condition with 2 distribution functions

@consts: wL, primL, hL, bL, bcL, wR, primR, hR, bR, bcR, bcU, bcD

KitBase.IB3F — Type

Initial & boundary condition with 3 distribution functions

@consts: wL, primL, h0L, h1L, h2L, bcL, EL, BL, lorenzL, wR, primR, h0R, h1R, h2R, bcR, ER, BR, lorenzR, bcU, bcD

KitBase.IB4F — Type

Initial & boundary condition with 4 distribution functions

@consts: wL, primL, h0L, h1L, h2L, h3L, bcL, EL, BL, lorenzL, wR, primR, h0R, h1R, h2R, h3R, bcR, ER, BR, lorenzR, bcU, bcD

KitBase.Interface1D — Type

1D cell interface with no distribution function

@vars: fw

Interface1D(w::AbstractArray)

KitBase.Interface1D1F — Type

1D cell interface with 1 distribution function

@vars: fw, ff

Interface1D1F(w::AbstractArray, f::AbstractArray)

KitBase.Interface1D2F — Type

1D cell interface with 2 distribution functions

@vars: fw, fh, fb

Interface1D2F(w::AbstractArray, f::AbstractArray)

KitBase.Interface1D3F — Type

1D cell interface with 3 distribution functions

@vars: fw, fh0, fh1, fh2, femL, femR,

deterministic: Interface1D3F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,1})

stochastic: Interface1D3F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,2})

KitBase.Interface1D4F — Type

1D cell interface with 4 distribution functions

@vars: fw, fh0, fh1, fh2, fh3, femL, femR,

deterministic: Interface1D4F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,1})

stochastic: Interface1D4F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,2})

KitBase.Interface2D — Type

2D cell interface with no distribution function

@vars: len, n, fw

Interface2D(L::Real, C::Real, S::Real, w::AbstractArray)

KitBase.Interface2D1F — Type

2D cell interface with 1 distribution function

@vars: len, n, fw, ff

Interface2D1F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

KitBase.Interface2D2F — Type

2D cell interface with 2 distribution functions

@vars: len, n, fw, fh, fb

Interface2D2F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

KitBase.MVSpace1D — Type

1D multi-component velocity space

@consts: u0, u1, nu, u, du, weights

KitBase.MVSpace2D — Type

2D multi-component velocity space

@consts: u0, u1, nu, v0, v1, nv, u, v, du, dv, weights

KitBase.MVSpace3D — Type

3D multi-component velocity space

@consts: u0, u1, nu, v0, v1, nv, w0, w1, nw, u, v, w, du, dv, dw, weights

KitBase.Mixture — Type

Multi-component gas property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne

KitBase.PSpace1D — Type

1D physical space with structured mesh

@consts: x0, x1, nx, x, dx

KitBase.PSpace2D — Type

2D Physical space with structured mesh

@consts: x0, x1, nx, y0, y1, ny, x, y, dx, dy

KitBase.Plasma1D — Type

1D plasma property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne, lD, rL, sol, χ, ν, Ap, An, D

KitBase.Plasma2D — Type

2D plasma property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne, lD, rL, sol, χ, ν, A1p, A1n, A2p, A2n, D1, D2

KitBase.Setup — Type

Computational setup

@consts: case, space, flux, collision, nSpecies, interpOrder, limiter, cfl, maxTime

KitBase.SolverSet — Type

Structure of solver setup

KitBase.UnstructMesh — Type

Physical space with unstructured mesh

@consts: nodes, cells

KitBase.VSpace1D — Type

1D velocity space

@consts: u0, u1, nu, u, du, weights

KitBase.VSpace2D — Type

2D velocity space

@consts: u0, u1, nu, v0, v1, nv, u, v, du, dv, weights

KitBase.VSpace3D — Type

3D velocity space

@consts: u0, u1, nu, v0, v1, nv, w0, w1, nw, u, v, w, du, dv, dw, weights