Initial and boundary conditions

A struct ib <: AbstractCondition defines the initial and boundary conditions of a simulation. It contains the values of conservative and primitive variables, and particle distribution functions at left and right (up and down) domain for both initial and boundary conditions. It is set this way to easily deal with discontinuous initial conditions.

KitBase.IB — Type

mutable struct IB{A,B,C} <: AbstractCondition
    wL::A
    primL::B
    bcL::B

    wR::A
    primR::B
    bcR::B

    bcU::B
    bcD::B

    vL::C
    vR::C
end

Initial & boundary condition with no distribution function

KitBase.IB1F — Type

mutable struct IB1F{A,B} <: AbstractCondition
    wL::A
    primL::A
    fL::B
    bcL::A

    wR::A
    primR::A
    fR::B
    bcR::A

    bcU::A
    bcD::A
end

Initial & boundary condition with 1 distribution function

KitBase.IB2F — Type

mutable struct IB2F{A,B} <: AbstractCondition
    wL::A
    primL::A
    hL::B
    bL::B
    bcL::A

    wR::A
    primR::A
    hR::B
    bR::B
    bcR::A

    bcU::A
    bcD::A
end

Initial & boundary condition with 2 distribution functions

KitBase.IB3F — Type

mutable struct IB3F{A,B,C,D} <: AbstractCondition
    wL::A
    primL::A
    h0L::B
    h1L::B
    h2L::B
    bcL::A
    EL::C
    BL::C
    lorenzL::D

    wR::A
    primR::A
    h0R::B
    h1R::B
    h2R::B
    bcR::A
    ER::C
    BR::C
    lorenzR::D

    bcU::A
    bcD::A
end

Initial & boundary condition with 3 distribution functions

KitBase.IB4F — Type

mutable struct IB4F{A,B,C,D} <: AbstractCondition
    wL::A
    primL::A
    h0L::B
    h1L::B
    h2L::B
    h3L::B
    bcL::A
    EL::C
    BL::C
    lorenzL::D

    wR::A
    primR::A
    h0R::B
    h1R::B
    h2R::B
    h3R::B
    bcR::A
    ER::C
    BR::C
    lorenzR::D

    bcU::A
    bcD::A
end

Initial & boundary condition with 4 distribution functions