Particle properties

A struct gas <: AbstractProperty defines the properties of particle model. It currently supports the following models:

scalar

gas-type molecule

plasma

KitBase.Scalar — Type

mutable struct Scalar{T, F} <: AbstractProperty

Fluid property for scalar conservation laws

Fields

a: Default: 1.0

μᵣ: Default: 1.0e-8

KitBase.Gas — Type

mutable struct Gas{T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11<:Integer, T12} <: AbstractGas

Gas property

Fields

Kn: Default: 0.01

Ma: Default: 0.0

Pr: Default: 1.0

K: Default: 2.0

γ: Default: 5 / 3

ω: Default: 0.81

αᵣ: Default: 1.0

ωᵣ: Default: 0.5

μᵣ: Default: refvhsvis(Kn, αᵣ, ωᵣ)

m: Default: 0.001

np: Default: 1000

fsm: Default: nothing

KitBase.Mixture — Type

struct Mixture{A, B, C, D, E, F, G, H, I} <: AbstractGas

Multi-component gas property

Fields

Kn: Default: 0.01

Ma: Default: 0.0

Pr: Default: 1.0

K: Default: 2.0

γ: Default: 5 / 3

mi: Default: 1.0

ni: Default: 0.5

me: Default: 0.5

ne: Default: 0.5

KitBase.Plasma1D — Type

struct Plasma1D{A, B, C, TD, E, F, G, H, I, J, TK, L, M, N, O, P} <: AbstractPlasma

1D plasma property

Fields

Kn: Default: 0.01

Ma: Default: 0.0

Pr: Default: 1.0

K: Default: 2.0

γ: Default: 5 / 3

mi: Default: 1.0

ni: Default: 0.5

me: Default: 0.5

ne: Default: 0.5

lD: Default: 0.01

rL: Default: 0.01

sol: Default: 100

χ: Default: 1

ν: Default: 1

Ap

An

D

KitBase.Plasma2D — Type

struct Plasma2D{A, B, C, D, E, F, G, H, I, J, TK, L, M, N, O, P} <: AbstractPlasma

2D plasma property

Fields

Kn: Default: 0.01

Ma: Default: 0.0

Pr: Default: 1.0

K: Default: 2

γ: Default: 5 / 3

mi: Default: 1.0

ni: Default: 0.5

me: Default: 0.5

ne: Default: 0.5

lD: Default: 0.01

rL: Default: 0.01

sol: Default: 100

χ: Default: 1

ν: Default: 1

A1p

A1n

A2p

A2n

D1

D2

The fields denote, for example:

Kn: reference Knudsen number

Ma: reference Mach number

Pr: reference Prandtl number

K: internal degree of freedom of molecule

γ: adiabatic index

ω: viscosity index

αᵣ: reference $\alpha$ in viscosity evaluation

ωᵣ: reference $\omega$ in viscosity evaluation

μᵣ: reference viscosity

m: mass of each particle

np: number of particles

The viscosity is evaluated the following hard-sphere model.

\[\mu = \mu_{ref} \left(\frac{T}{T_{ref}}\right)^{\omega}\]

\[\mu_{ref}=\frac{5(\alpha+1)(\alpha+2) \sqrt{\pi}}{4 \alpha(5-2 \omega)(7-2 \omega)} Kn_{ref}\]