Particle velocity space

A struct vSpace <: AbstractSetup defines the particle velocity setup of a simulation. Structs for 1-3 dimensional particle velocity space are built. Note that these structs work for both single-component and multi-component substances.

KitBase.VSpace1D — Type

struct VSpace1D{TR, TI, TA, TB} <: AbstractVelocitySpace1D

1D velocity space

Fields

u0

u1

nu

u

du

weights

KitBase.VSpace2D — Type

struct VSpace2D{TR, TI, TA} <: AbstractVelocitySpace2D

2D velocity space

Fields

u0

u1

nu

v0

v1

nv

u

v

du

dv

weights

KitBase.VSpace3D — Type

struct VSpace3D{TR, TI, TA} <: KitBase.AbstractVelocitySpace3D

3D velocity space

Fields

u0

u1

nu

v0

v1

nv

w0

w1

nw

u

v

w

du

dv

dw

weights

It contains

u0 (v0, w0): location of starting point

u1 (v1, w1): location of ending point

nu (nv, nw): number of cells in one direction

u (v, w): locations of middle points of all cells

du (dv, dw): intervals of all cell points

weights: quadrature weights for numerical integral

Note that the one-dimensional velocity space can be used to handle 1-3 dimensional unstructured topology as well.

For the simulation cases where no phase-space evolution is involved, vSpace can be set as nothing directly.

vSpace = nothing