Theory

Kinetic.prim_conserve — Function

Transform primitive -> conservative variables

Kinetic.conserve_prim — Function

Transform conservative -> primitive variables

scalar: pseudo primitive vector for scalar conservation laws

vector: primitive vector for Euler, Navier-Stokes and extended equations

Missing docstring.

Missing docstring for mixture_prim_conserve. Check Documenter's build log for details.

Missing docstring.

Missing docstring for mixture_conserve_prim. Check Documenter's build log for details.

Kinetic.em_coefficients — Function

Calculate electromagnetic coeffcients in hyperbolic Maxwell's equations

Kinetic.advection_flux — Function

Theoretical flux of linear advection equation

Kinetic.burgers_flux — Function

Theoretical flux of Burgers' equation

Kinetic.euler_flux — Function

Theoretical fluxes of Euler Equations

@return: flux tuple

Kinetic.euler_jacobi — Function

Flux Jacobian of Euler Equations

Kinetic.gauss_moments — Function

Calculate moments of Gaussian distribution G = (λ / π)^(D / 2) * exp[-λ(c^2 + ξ^2)]

internality: gauss_moments(prim::AbstractArray{<:Real,1}, inK::Real)

no internality: gauss_moments(prim::AbstractArray{<:Real,1})

Kinetic.mixture_gauss_moments — Function

Calculate moments of Gaussian distribution in multi-component gas

mixture_gauss_moments(prim::AbstractArray{<:Real,2}, inK::Real)

Kinetic.moments_conserve — Function

Calculate conservative moments of particle distribution

moments_conserve(Mu::OffsetArray{<:AbstractFloat,1}, alpha::Int)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mxi::OffsetArray{<:Real,1}, alpha::Int, delta::Int)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mv::OffsetArray{<:Real,1}, Mw::OffsetArray{<:Real,1}, alpha::Int, beta::Int, delta::Int)

Kinetic.mixture_moments_conserve — Function

Calculate conservative moments of particle distribution in multi-component gas

Kinetic.pdf_slope — Function

Calculate slope of particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

pdf_slope(u::Real, Δ::Real)

pdf_slope(prim::AbstractArray{<:Real,1}, sw::AbstractArray{<:Real,1}, inK::Real)

Kinetic.mixture_pdf_slope — Function

Calculate slope of multi-component particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

mixture_pdf_slope(prim::AbstractArray{<:Real,2}, sw::AbstractArray{<:Real,2}, inK::Real)

Kinetic.moments_conserve_slope — Function

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

Kinetic.mixture_moments_conserve_slope — Function

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

Kinetic.discrete_moments — Function

Discrete moments of particle distribution

discrete_moments(f, ω): direct quadrature

discrete_moments(f, u, ω, n): velocity moments

Kinetic.stress — Function

Calculate stress tensor from particle distribution function

Kinetic.heat_flux — Function

Calculate heat flux from particle distribution function

Kinetic.maxwellian — Function

Maxwellian in discrete form

@arg: particle velocity quadrature points

@arg: density, velocity and inverse of temperature

@return: Maxwellian distribution function

Missing docstring.

Missing docstring for mixture_maxwellian. Check Documenter's build log for details.

Kinetic.shakhov — Function

Shakhov non-equilibrium part

@arg: particle velocity quadrature points

@arg: discrete Maxwellian

@arg: primitive variables, Prandtl number, heat flux, inner degree of freedom

Kinetic.reduce_distribution — Function

Reduced distribution function

@arg : particle distribution function with full velocity space

@arg : quadrature weights with reduced velocity setting (v & w by default)

Missing docstring.

Missing docstring for full_distribution. Check Documenter's build log for details.

Kinetic.ref_vhs_vis — Function

Calculate reference viscosity

variable hard sphere (VHS) model

Kinetic.vhs_collision_time — Function

Calculate collision time

variable hard sphere (VHS) model

Kinetic.aap_hs_collision_time — Function

Calculate mixture collision time from AAP model

aap_hs_collision_time( prim::AbstractArray{<:Real,2}, mi::Real, ni::Real, me::Real, ne::Real, kn::Real, )

Kinetic.aap_hs_prim — Function

Calculate mixture primitive variables from AAP model

tested √

aap_hs_prim( prim::AbstractArray{<:Real,2}, tau::AbstractArray{<:Real,1}, mi::Real, ni::Real, me::Real, ne::Real, kn::Real, )

Kinetic.aap_hs_diffeq! — Function

Source term of AAP model in DifferentialEquations.jl

aap_hs_diffeq!(du, u, p, t)

Kinetic.shift_pdf! — Function

Shift distribution function by external force

shift_pdf!( f::AbstractArray{<:AbstractFloat,1}, a::Real, du::AbstractFloat, dt::Real, )

shift_pdf!( f::AbstractArray{<:AbstractFloat,2}, a::AbstractArray{<:Real,1}, du::AbstractArray{<:AbstractFloat,1}, dt::Real, )

Kinetic.hs_boltz_kn — Function

Calculate effective Knudsen number for fast spectral method

hard sphere (HS) model

Kinetic.kernel_mode — Function

Calculate collision kernel for fast spectral method

Kinetic.boltzmann_fft — Function

Calculate collision operator with FFT-based fast spectral method

Missing docstring.

Missing docstring for boltzmann_fft!. Check Documenter's build log for details.

Kinetic.heat_capacity_ratio — Function

Calculate heat capacity ratio

Kinetic.sound_speed — Function

Calculate speed of sound