Theory

KitBase.prim_conserve — Function

Transform primitive -> conservative variables

prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,1}}

prim_conserve(ρ, U, λ, γ)

prim_conserve(ρ, U, V, λ, γ)

prim_conserve(ρ, U, V, W, λ, γ)

KitBase.conserve_prim — Function

Transform conservative -> primitive variables

scalar: pseudo primitive vector for scalar conservation laws

conserve_prim(u)

conserve_prim(u, a)

vector: primitive vector for Euler, Navier-Stokes and extended equations

conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,1}}

conserve_prim(ρ, M, E, γ)

conserve_prim(ρ, MX, MY, E, γ)

KitBase.mixture_prim_conserve — Function

Transform multi-component primitive -> conservative variables

mixture_prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,2}}

KitBase.mixture_conserve_prim — Function

Transform multi-component conservative -> primitive variables

mixture_conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,2}}

KitBase.em_coefficients — Function

Calculate electromagnetic coeffcients in hyperbolic Maxwell's equations

em_coefficients(
    prim::X,
    E::Y,
    B::Z,
    mr,
    lD,
    rL,
    dt,
) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:Real,1}}

KitBase.advection_flux — Function

Theoretical flux of linear advection equation

advection_flux(u, a)

KitBase.burgers_flux — Function

Theoretical flux of Burgers' equation

burgers_flux(u)

KitBase.euler_flux — Function

Theoretical fluxes of Euler Equations

euler_flux(w::A, γ; frame = :cartesian::Symbol) where {A<:AbstractArray{<:Real,1}}

@return: flux tuple

KitBase.euler_jacobi — Function

Flux Jacobian of Euler Equations

euler_jacobi(w::T, γ) where {T<:AbstractArray{<:Real,1}}

@return: Jacobian matrix A

KitBase.gauss_moments — Function

Calculate moments of Gaussian distribution G = (λ / π)^(D / 2) * exp[-λ(c^2 + ξ^2)]

internality: gauss_moments(prim::T) where {T<:AbstractArray{<:Real,1}}

no internality: gauss_moments(prim::T, inK) where {T<:AbstractArray{<:Real,1}}

KitBase.mixture_gauss_moments — Function

Calculate moments of Gaussian distribution in multi-component gas

mixture_gauss_moments(prim::T, inK) where {T<:AbstractArray{<:Real,2}}

KitBase.moments_conserve — Function

Calculate conservative moments of particle distribution

moments_conserve(Mu::OffsetArray{<:AbstractFloat,1}, alpha::Int)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mxi::OffsetArray{<:Real,1}, alpha::Int, delta::Int)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mv::OffsetArray{<:Real,1}, Mw::OffsetArray{<:Real,1}, alpha::Int, beta::Int, delta::Int)

KitBase.mixture_moments_conserve — Function

Calculate conservative moments of particle distribution in multi-component gas

mixture_moments_conserve(
    Mu::T,
    Mxi::T,
    alpha::I,
    delta::I,
) where {T<:OffsetArray{<:AbstractFloat,2},I<:Int}

function mixture_moments_conserve(
    Mu::T,
    Mv::T,
    Mw::T,
    alpha::I,
    beta::I,
    delta::I,
) where {T<:OffsetArray{<:AbstractFloat,2},I<:Int}

KitBase.pdf_slope — Function

Calculate slope of particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

pdf_slope(u, Δ)

pdf_slope(prim::A, sw::B, inK) where {A<:AbstractArray{<:Real,1},B<:AbstractArray{<:Real,1}}

KitBase.mixture_pdf_slope — Function

Calculate slope of multi-component particle distribution function, assuming a = a1 + u * a2 + 0.5 * u^2 * a3

mixture_pdf_slope(prim::X, sw::Y, inK) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,2}}

KitBase.moments_conserve_slope — Function

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.mixture_moments_conserve_slope — Function

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.discrete_moments — Function

Discrete moments of particle distribution

discrete_moments(f, ω): direct quadrature

discrete_moments(f, u, ω, n): velocity moments

KitBase.stress — Function

Calculate stress tensor from particle distribution function

KitBase.heat_flux — Function

Calculate heat flux from particle distribution function

KitBase.maxwellian — Function

Maxwellian in discrete form

@args: particle velocity quadrature points

@args: density, velocity and inverse of temperature

@return: Maxwellian distribution function

KitBase.mixture_maxwellian — Function

Multi-component Maxwellian in discrete form

mixture_maxwellian(u::X, prim::Y) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,2}}

mixture_maxwellian(
    u::X,
    v::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,3},Y<:AbstractArray{<:Real,2}}

mixture_maxwellian(
    u::X,
    v::X,
    w::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,4},Y<:AbstractArray{<:Real,2}}

KitBase.shakhov — Function

Shakhov non-equilibrium part

@arg: particle velocity quadrature points

@arg: discrete Maxwellian

@arg: primitive variables, Prandtl number, heat flux, inner degree of freedom

KitBase.reduce_distribution — Function

Reduced distribution function

@arg : particle distribution function with full velocity space

@arg : quadrature weights with reduced velocity setting (v & w by default)

KitBase.full_distribution — Function

Recover full distribution function from reduced ones

@arg h & b : reduced particle distribution function with 1D velocity space

@arg u : quadrature nodes in 1D velocity space

@arg weights : quadrature weights in 1D velocity space

@arg v & w : quadrature nodes in the rest velocity space (with 3D setting)

@return f : particle distribution function with 3D velocity space

KitBase.ref_vhs_vis — Function

Calculate reference viscosity

variable hard sphere (VHS) model

KitBase.vhs_collision_time — Function

Calculate collision time

variable hard sphere (VHS) model

KitBase.aap_hs_collision_time — Function

Calculate mixture collision time from AAP model

aap_hs_collision_time( prim::AbstractArray{<:Real,2}, mi::Real, ni::Real, me::Real, ne::Real, kn::Real, )

KitBase.aap_hs_prim — Function

Calculate mixture primitive variables from AAP model

tested √

aap_hs_prim( prim::AbstractArray{<:Real,2}, tau::AbstractArray{<:Real,1}, mi::Real, ni::Real, me::Real, ne::Real, kn::Real, )

KitBase.aap_hs_diffeq! — Function

Source term of AAP model in DifferentialEquations.jl

aap_hs_diffeq!(du, u, p, t)

KitBase.shift_pdf! — Function

Shift distribution function by external force

shift_pdf!(
    f::T,
    a,
    du,
    dt,
) where {T<:AbstractArray{<:AbstractFloat,1}}

shift_pdf!(
    f::X,
    a::Y,
    du::Z,
    dt,
) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:AbstractFloat,1}}

KitBase.hs_boltz_kn — Function

Calculate effective Knudsen number for fast spectral method

hard sphere (HS) model

KitBase.kernel_mode — Function

Calculate collision kernel for fast spectral method

KitBase.boltzmann_fft — Function

Calculate collision operator with FFT-based fast spectral method

KitBase.boltzmann_fft! — Function

Calculate collision operator with FFT-based fast spectral method

KitBase.heat_capacity_ratio — Function

Calculate heat capacity ratio

KitBase.sound_speed — Function

Calculate speed of sound