Theory

KitBase.prim_conserve — Function

prim_conserve(prim::AV{T}, γ) where {T<:Real}
prim_conserve(ρ, U, λ, γ)
prim_conserve(ρ, U, V, λ, γ)
prim_conserve(ρ, U, V, W, λ, γ)

Transform primitive -> conservative variables

KitBase.conserve_prim — Function

scalar: pseudo primitive vector for scalar conservation laws

conserve_prim(u)

conserve_prim(u, a)

vector: primitive vector for Euler, Navier-Stokes and extended equations

conserve_prim(W::T, γ) where {T<:AA{<:Real,1}}

conserve_prim(ρ, M, E, γ)

conserve_prim(ρ, MX, MY, E, γ)

Transform conservative -> primitive variables

KitBase.mixture_prim_conserve — Function

mixture_prim_conserve(prim::AM{T}, γ) where {T<:Real}

Transform multi-component primitive -> conservative variables

KitBase.mixture_conserve_prim — Function

mixture_conserve_prim(W::T, γ) where {T<:AA{<:Real,2}}

Transform multi-component conservative -> primitive variables

KitBase.em_coefficients — Function

em_coefficients(prim, E, B, mr, ld, rL, dt)

Calculate electromagnetic coeffcients in hyperbolic Maxwell's equations

KitBase.advection_flux — Function

advection_flux(u, a)

Theoretical flux of linear advection equation

KitBase.burgers_flux — Function

burgers_flux(u)

Theoretical flux of Burgers' equation

KitBase.euler_flux — Function

euler_flux(w::AV{T}, γ; frame = :cartesian::Symbol) where {T<:Real}

Theoretical fluxes of Euler Equations

@return: flux tuple

KitBase.euler_jacobi — Function

euler_jacobi(w::AV{T}, γ) where {T<:Real}

Flux Jacobian of Euler Equations

@return: Jacobian matrix A

KitBase.gauss_moments — Function

Calculate moments of Gaussian distribution G = (λ / π)^(D / 2) * exp[-λ(c^2 + ξ^2)]

internality: gauss_moments(prim::T) where {T<:AA{<:Real,1}}

no internality: gauss_moments(prim::T, inK) where {T<:AA{<:Real,1}}

KitBase.mixture_gauss_moments — Function

mixture_gauss_moments(prim::AM{T}, inK) where {T<:Real}

Calculate moments of Gaussian distribution in multi-component gas

KitBase.moments_conserve — Function

moments_conserve(Mu::OffsetArray{<:FN,1}, alpha::Integer)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mxi::OffsetArray{<:Real,1},
    alpha::Integer, delta::Integer)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mv::OffsetArray{<:Real,1},
    Mw::OffsetArray{<:Real,1}, alpha::Integer, beta::Integer, delta::Integer)

Calculate conservative moments of particle distribution

KitBase.mixture_moments_conserve — Function

mixture_moments_conserve(
    Mu::T,
    Mxi::T,
    alpha::I,
    delta::I,
) where {T<:OffsetArray{<:Real,2},I<:Int}

function mixture_moments_conserve(
    Mu::T,
    Mv::T,
    Mw::T,
    alpha::I,
    beta::I,
    delta::I,
) where {T<:OffsetArray{<:Real,2},I<:Int}

Calculate conservative moments of particle distribution in multi-component gas

KitBase.pdf_slope — Function

pdf_slope(u, Δ)
pdf_slope(prim, sw, inK)

Calculate slope of particle distribution function a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.mixture_pdf_slope — Function

mixture_pdf_slope(prim, sw, inK)

Calculate slope of multi-component particle distribution function a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.moments_conserve_slope — Function

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.mixture_moments_conserve_slope — Function

mixturemomentsconserve_slope( a::X, Mu::Y, Mxi::Y, alpha::I, ) where {X<:AA{<:Real,2},Y<:OffsetArray{<:Real,2},I<:Integer}

mixturemomentsconserve_slope( a::X, Mu::Y, Mv::Y, Mxi::Y, alpha::I, beta::I, ) where {X<:AA{<:Real,2},Y<:OffsetArray{<:Real,2},I<:Integer}

mixturemomentsconserve_slope( a::X, Mu::Y, Mv::Y, Mw::Y, alpha::I, beta::I, delta::I, ) where {X<:AA{<:Real,2},Y<:OffsetArray{<:Real,2},I<:Integer}

Calculate slope-related conservative moments under the assumption a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.discrete_moments — Function

* direct quadrature
discrete_moments(f, ω)

* velocity moments
discrete_moments(f, u, ω, n)

Discrete moments of particle distribution

KitBase.stress — Function

Calculate stress tensor from particle distribution function

KitBase.heat_flux — Function

heat_flux(f, prim, u, ω)
heat_flux(h, b, prim, u, ω)
heat_flux(h, b, r, prim, u, ω)
heat_flux(f, prim, u, v, ω)
heat_flux(h, b, prim, u, v, ω)
heat_flux(h, b, r, prim, u, v, ω)
heat_flux(f, prim, u, v, w, ω)

Calculate heat flux from particle distribution function

Multiple dispatch doesn't consider unstructured multi-dimensional velocity space. In that case a new method needs to be defined.

KitBase.maxwellian — Function

maxwellian(u, ρ, U, λ)
maxwellian(u, prim)
maxwellian(u, v, ρ, U, V, λ)
maxwellian(u, v, prim)
maxwellian(u, v, w, ρ, U, V, W, λ)
maxwellian(u, v, w, prim)

Maxwellian in discrete form

@args: particle velocity quadrature points

@args: density, velocity and inverse of temperature

@return: Maxwellian distribution function

KitBase.mixture_maxwellian — Function

mixture_maxwellian(u, prim)
mixture_maxwellian(u, v, prim)
mixture_maxwellian(u, v, w, prim)

Multi-component Maxwellian in discrete form

KitBase.shakhov — Function

shakhov(u, M, q, prim, Pr)
shakhov(u, H, B, q, prim, Pr, K)
shakhov(u, v, M, q, prim, Pr)
shakhov(u, v, H, B, q, prim, Pr, K)
shakhov(u, v, w, M, q, prim, Pr)

Shakhov non-equilibrium part

@arg: particle velocity quadrature points

@arg: discrete Maxwellian

@arg: primitive variables, Prandtl number, heat flux, inner degree of freedom

KitBase.reduce_distribution — Function

reduce_distribution(f, weights, dim)
reduce_distribution(f, v, w, weights, dim)

Reduced distribution function

@arg : particle distribution function with full velocity space

@arg : quadrature weights with reduced velocity setting (v & w by default)

KitBase.full_distribution — Function

full_distribution(h, b, u, weights, v, w, ρ, γ)
full_distribution(h, b, u, weights, v, w, prim, γ)

Recover full distribution function from reduced ones

@arg h & b : reduced particle distribution function with 1D velocity space

@arg u : quadrature nodes in 1D velocity space

@arg weights : quadrature weights in 1D velocity space

@arg v & w : quadrature nodes in the rest velocity space (with 3D setting)

@return f : particle distribution function with 3D velocity space

KitBase.ref_vhs_vis — Function

ref_vhs_vis(Kn, alpha, omega)

Calculate reference viscosity with variable hard sphere (VHS) model

KitBase.vhs_collision_time — Function

vhs_collision_time(ρ, λ, μᵣ, ω)
vhs_collision_time(prim::AV{T}, muRef, omega) where {T<:Real}

Calculate collision time with variable hard sphere (VHS) model

KitBase.aap_hs_collision_time — Function

aap_hs_collision_time(
    prim::AA{<:Real,2},
    mi::Real,
    ni::Real,
    me::Real,
    ne::Real,
    kn::Real,
)

Calculate mixture collision time from AAP model

KitBase.aap_hs_prim — Function

aap_hs_prim(
    prim::AA{<:Real,2},
    tau::AA{<:Real,1},
    mi::Real,
    ni::Real,
    me::Real,
    ne::Real,
    kn::Real,
)

Calculate mixture primitive variables from AAP model

KitBase.aap_hs_diffeq! — Function

aap_hs_diffeq!(du, u, p, t)

Source term of AAP model in DifferentialEquations.jl

KitBase.shift_pdf! — Function

shift_pdf!(f, a, du, dt)

Shift distribution function by external force

KitBase.hs_boltz_kn — Function

hs_boltz_kn(mu_ref, alpha)

Calculate effective Knudsen number for fast spectral method with hard sphere (HS) model

KitBase.kernel_mode — Function

kernel_mode(
    M::TI,
    umax::TR,
    vmax::TR,
    du::TR,
    dv::TR,
    unum::TI,
    vnum::TI;
    quad_num = 64::TI,
) where {TI<:Integer,TR<:Real}

kernel_mode(
    M::I,
    umax::R,
    vmax::R,
    wmax::R,
    du::R,
    dv::R,
    dw::R,
    unum::I,
    vnum::I,
    wnum::I,
    alpha::R;
    quad_num = 64::I,
) where {I<:Integer,R<:Real}

Calculate collision kernel for fast spectral method

KitBase.boltzmann_fft — Function

boltzmann_fft(
    f::X,
    Kn,
    M::I,
    ϕ::Y,
    ψ::Y,
    phipsi::Z,
) where {
    X<:AA{<:Real,3},
    Y<:AA{<:Real,4},
    Z<:AA{<:Real,3},
    I<:Integer,
}

Calculate collision operator with FFT-based fast spectral method

KitBase.boltzmann_fft! — Function

boltzmann_fft!(
    Q::T1,
    f::T2,
    Kn::TR,
    M::TI,
    ϕ::TY,
    ψ::TY,
    phipsi::TZ,
) where {
    T1<:AA{<:Real,3},
    T2<:AA{<:Real,3},
    TR<:Real,
    TI<:Integer,
    TY<:AA{<:Real,4},
    TZ<:AA{<:Real,3},
}

Calculate collision operator with FFT-based fast spectral method

KitBase.heat_capacity_ratio — Function

heat_capacity_ratio(K, D)
heat_capacity_ratio(K, Nr, D)

Calculate heat capacity ratio

KitBase.sound_speed — Function

sound_speed(λ, γ)
sound_speed(prim, γ)

Calculate speed of sound