Theory

KitBase.prim_conserve — Function

prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,1}}

prim_conserve(ρ, U, λ, γ)

prim_conserve(ρ, U, V, λ, γ)

prim_conserve(ρ, U, V, W, λ, γ)

Transform primitive -> conservative variables

KitBase.conserve_prim — Function

scalar: pseudo primitive vector for scalar conservation laws

conserve_prim(u)

conserve_prim(u, a)

vector: primitive vector for Euler, Navier-Stokes and extended equations

conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,1}}

conserve_prim(ρ, M, E, γ)

conserve_prim(ρ, MX, MY, E, γ)

Transform conservative -> primitive variables

KitBase.mixture_prim_conserve — Function

mixture_prim_conserve(prim::T, γ) where {T<:AbstractArray{<:Real,2}}

Transform multi-component primitive -> conservative variables

KitBase.mixture_conserve_prim — Function

mixture_conserve_prim(W::T, γ) where {T<:AbstractArray{<:Real,2}}

Transform multi-component conservative -> primitive variables

KitBase.em_coefficients — Function

em_coefficients(
    prim::X,
    E::Y,
    B::Z,
    mr,
    lD,
    rL,
    dt,
) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:Real,1}}

Calculate electromagnetic coeffcients in hyperbolic Maxwell's equations

KitBase.advection_flux — Function

advection_flux(u, a)

Theoretical flux of linear advection equation

KitBase.burgers_flux — Function

burgers_flux(u)

Theoretical flux of Burgers' equation

KitBase.euler_flux — Function

euler_flux(w::A, γ; frame = :cartesian::Symbol) where {A<:AbstractArray{<:Real,1}}

Theoretical fluxes of Euler Equations

@return: flux tuple

KitBase.euler_jacobi — Function

euler_jacobi(w::T, γ) where {T<:AbstractArray{<:Real,1}}

Flux Jacobian of Euler Equations

@return: Jacobian matrix A

KitBase.gauss_moments — Function

Calculate moments of Gaussian distribution G = (λ / π)^(D / 2) * exp[-λ(c^2 + ξ^2)]

internality: gauss_moments(prim::T) where {T<:AbstractArray{<:Real,1}}

no internality: gauss_moments(prim::T, inK) where {T<:AbstractArray{<:Real,1}}

KitBase.mixture_gauss_moments — Function

mixture_gauss_moments(prim::T, inK) where {T<:AbstractArray{<:Real,2}}

Calculate moments of Gaussian distribution in multi-component gas

KitBase.moments_conserve — Function

moments_conserve(Mu::OffsetArray{<:AbstractFloat,1}, alpha::Integer)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mxi::OffsetArray{<:Real,1},
    alpha::Integer, delta::Integer)

moments_conserve(Mu::OffsetArray{<:Real,1}, Mv::OffsetArray{<:Real,1},
    Mw::OffsetArray{<:Real,1}, alpha::Integer, beta::Integer, delta::Integer)

Calculate conservative moments of particle distribution

KitBase.mixture_moments_conserve — Function

mixture_moments_conserve(
    Mu::T,
    Mxi::T,
    alpha::I,
    delta::I,
) where {T<:OffsetArray{<:AbstractFloat,2},I<:Int}

function mixture_moments_conserve(
    Mu::T,
    Mv::T,
    Mw::T,
    alpha::I,
    beta::I,
    delta::I,
) where {T<:OffsetArray{<:AbstractFloat,2},I<:Int}

Calculate conservative moments of particle distribution in multi-component gas

KitBase.pdf_slope — Function

pdf_slope(u, Δ)

pdf_slope(prim::A, sw::B, inK) where {A<:AbstractArray{<:Real,1},B<:AbstractArray{<:Real,1}}

Calculate slope of particle distribution function a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.mixture_pdf_slope — Function

mixture_pdf_slope(prim::X, sw::Y, inK) where {X<:AbstractArray{<:Real,2},Y<:AbstractArray{<:Real,2}}

Calculate slope of multi-component particle distribution function a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.moments_conserve_slope — Function

Calculate slope-related conservative moments a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.mixture_moments_conserve_slope — Function

mixturemomentsconserve_slope( a::X, Mu::Y, Mxi::Y, alpha::I, ) where {X<:AbstractArray{<:Real,2},Y<:OffsetArray{<:Real,2},I<:Integer}

mixturemomentsconserve_slope( a::X, Mu::Y, Mv::Y, Mxi::Y, alpha::I, beta::I, ) where {X<:AbstractArray{<:Real,2},Y<:OffsetArray{<:Real,2},I<:Integer}

mixturemomentsconserve_slope( a::X, Mu::Y, Mv::Y, Mw::Y, alpha::I, beta::I, delta::I, ) where {X<:AbstractArray{<:Real,2},Y<:OffsetArray{<:Real,2},I<:Integer}

Calculate slope-related conservative moments under the assumption a = a1 + u * a2 + 0.5 * u^2 * a3

KitBase.discrete_moments — Function

Discrete moments of particle distribution

discrete_moments(f, ω): direct quadrature

discrete_moments(f, u, ω, n): velocity moments

KitBase.stress — Function

Calculate stress tensor from particle distribution function

KitBase.heat_flux — Function

Calculate heat flux from particle distribution function

KitBase.maxwellian — Function

maxwellian(u::T, ρ, U, λ) where {T<:AbstractArray{<:AbstractFloat,1}}

maxwellian(
    u::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,1},Y<:AbstractArray{<:Real,1}}

maxwellian(u::T, v::T, ρ, U, V, λ) where {T<:AbstractArray{<:AbstractFloat,2}}

maxwellian(
    u::X,
    v::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,1}}

maxwellian(u::T, v::T, w::T, ρ, U, V, W, λ) where {T<:AbstractArray{<:AbstractFloat,3}}

maxwellian(
    u::X,
    v::X,
    w::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,3},Y<:AbstractArray{<:Real,1}}

Maxwellian in discrete form

@args: particle velocity quadrature points

@args: density, velocity and inverse of temperature

@return: Maxwellian distribution function

KitBase.mixture_maxwellian — Function

mixture_maxwellian(u::X, prim::Y) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,2}}

mixture_maxwellian(
    u::X,
    v::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,3},Y<:AbstractArray{<:Real,2}}

mixture_maxwellian(
    u::X,
    v::X,
    w::X,
    prim::Y,
) where {X<:AbstractArray{<:AbstractFloat,4},Y<:AbstractArray{<:Real,2}}

Multi-component Maxwellian in discrete form

KitBase.shakhov — Function

shakhov(
    u::X,
    M::Y,
    q,
    prim::Z,
    Pr,
) where {
    X<:AbstractArray{<:AbstractFloat,1},
    Y<:AbstractArray{<:AbstractFloat,1},
    Z<:AbstractArray{<:Real,1},
}

shakhov(
    u::T,
    H::X,
    B::X,
    q,
    prim::Y,
    Pr,
    K,
) where {
    T<:AbstractArray{<:AbstractFloat,1},
    X<:AbstractArray{<:AbstractFloat,1},
    Y<:AbstractArray{<:Real,1},
}

shakhov(
    u::T,
    v::T,
    M::X,
    q::Y,
    prim::Z,
    Pr,
) where {
    T<:AbstractArray{<:AbstractFloat,2},
    X<:AbstractArray{<:AbstractFloat,2},
    Y<:AbstractArray{<:AbstractFloat,1},
    Z<:AbstractArray{<:Real,1},
}

shakhov(
    u::T,
    v::T,
    H::X,
    B::X,
    q::Y,
    prim::Z,
    Pr,
    K,
) where {
    T<:AbstractArray{<:AbstractFloat,2},
    X<:AbstractArray{<:AbstractFloat,2},
    Y<:AbstractArray{<:Real,1},
    Z<:AbstractArray{<:Real,1},
}

shakhov(
    u::T,
    v::T,
    w::T,
    M::X,
    q::Y,
    prim::Z,
    Pr,
) where {
    T<:AbstractArray{<:AbstractFloat,3},
    X<:AbstractArray{<:AbstractFloat,3},
    Y<:AbstractArray{<:Real,1},
    Z<:AbstractArray{<:Real,1},
}

Shakhov non-equilibrium part

@arg: particle velocity quadrature points

@arg: discrete Maxwellian

@arg: primitive variables, Prandtl number, heat flux, inner degree of freedom

KitBase.reduce_distribution — Function

reduce_distribution(
    f::X,
    weights::Y,
    dim = 1,
) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:AbstractFloat,1}}

reduce_distribution(
    f::X,
    weights::Y,
    dim = 1,
) where {X<:AbstractArray{<:AbstractFloat,3},Y<:AbstractArray{<:AbstractFloat,2}}

Reduced distribution function

@arg : particle distribution function with full velocity space

@arg : quadrature weights with reduced velocity setting (v & w by default)

KitBase.full_distribution — Function

full_distribution(
    h::X,
    b::X,
    u::Y,
    weights::Y,
    v::Z,
    w::Z,
    ρ,
    γ = 5 / 3,
) where {
    X<:AbstractArray{<:AbstractFloat,1},
    Y<:AbstractArray{<:AbstractFloat,1},
    Z<:AbstractArray{<:AbstractFloat,3},
}

full_distribution(
    h::X,
    b::X,
    u::Y,
    weights::Y,
    v::Z,
    w::Z,
    prim::A,
    γ = 5 / 3,
) where {
    X<:AbstractArray{<:AbstractFloat,1},
    Y<:AbstractArray{<:AbstractFloat,1},
    Z<:AbstractArray{<:AbstractFloat,3},
    A<:AbstractArray{<:Real,1},
}

Recover full distribution function from reduced ones

@arg h & b : reduced particle distribution function with 1D velocity space

@arg u : quadrature nodes in 1D velocity space

@arg weights : quadrature weights in 1D velocity space

@arg v & w : quadrature nodes in the rest velocity space (with 3D setting)

@return f : particle distribution function with 3D velocity space

KitBase.ref_vhs_vis — Function

ref_vhs_vis(Kn, alpha, omega)

Calculate reference viscosity with variable hard sphere (VHS) model

KitBase.vhs_collision_time — Function

vhs_collision_time(prim::T, muRef, omega) where {T<:AbstractArray{<:Real,1}}

Calculate collision time with variable hard sphere (VHS) model

KitBase.aap_hs_collision_time — Function

aap_hs_collision_time(
    prim::AbstractArray{<:Real,2},
    mi::Real,
    ni::Real,
    me::Real,
    ne::Real,
    kn::Real,
)

Calculate mixture collision time from AAP model

KitBase.aap_hs_prim — Function

aap_hs_prim(
    prim::AbstractArray{<:Real,2},
    tau::AbstractArray{<:Real,1},
    mi::Real,
    ni::Real,
    me::Real,
    ne::Real,
    kn::Real,
)

Calculate mixture primitive variables from AAP model

KitBase.aap_hs_diffeq! — Function

aap_hs_diffeq!(du, u, p, t)

Source term of AAP model in DifferentialEquations.jl

KitBase.shift_pdf! — Function

shift_pdf!(
    f::T,
    a,
    du,
    dt,
) where {T<:AbstractArray{<:AbstractFloat,1}}

shift_pdf!(
    f::X,
    a::Y,
    du::Z,
    dt,
) where {X<:AbstractArray{<:AbstractFloat,2},Y<:AbstractArray{<:Real,1},Z<:AbstractArray{<:AbstractFloat,1}}

Shift distribution function by external force

KitBase.hs_boltz_kn — Function

hs_boltz_kn(mu_ref, alpha)

Calculate effective Knudsen number for fast spectral method with hard sphere (HS) model

KitBase.kernel_mode — Function

kernel_mode(
    M::I,
    umax::R,
    vmax::R,
    wmax::R,
    du::R,
    dv::R,
    dw::R,
    unum::I,
    vnum::I,
    wnum::I,
    alpha::R;
    quad_num = 64::I,
) where {I<:Integer,R<:Real}

Calculate collision kernel for fast spectral method

KitBase.boltzmann_fft — Function

boltzmann_fft(
    f::X,
    Kn,
    M::I,
    ϕ::Y,
    ψ::Y,
    phipsi::Z,
) where {
    X<:AbstractArray{<:AbstractFloat,3},
    Y<:AbstractArray{<:AbstractFloat,4},
    Z<:AbstractArray{<:AbstractFloat,3},
    I<:Integer,
}

Calculate collision operator with FFT-based fast spectral method

KitBase.boltzmann_fft! — Function

boltzmann_fft!(
    Q::X,
    f::X,
    Kn,
    M::I,
    ϕ::Y,
    ψ::Y,
    phipsi::Z,
) where {
    X<:AbstractArray{<:AbstractFloat,3},
    Y<:AbstractArray{<:AbstractFloat,4},
    Z<:AbstractArray{<:AbstractFloat,3},
    I<:Integer,
}

Calculate collision operator with FFT-based fast spectral method

KitBase.heat_capacity_ratio — Function

heat_capacity_ratio(K, D::T) where {T<:Integer}

Calculate heat capacity ratio

KitBase.sound_speed — Function

sound_speed(λ::Real, γ::Real)
sound_speed(prim::T, γ) where {T<:AbstractArray{<:Real,1}}
sound_speed(prim::T, γ) where {T<:AbstractArray{<:Real,2}}

Calculate speed of sound