Data Structure

Kinetic.ControlVolume1D — Type

1D control volume with no distribution function

ControlVolume1D(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray)

@vars: x, dx, w, prim, sw

Kinetic.ControlVolume1D1F — Type

1D control volume with 1 distribution function

ControlVolume1D1F(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray, F::AbstractArray)

@vars: x, dx, w, prim, sw, f, sf

Kinetic.ControlVolume1D2F — Type

1D control volume with 2 distribution functions

ControlVolume1D2F(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray, H::AbstractArray, B::AbstractArray)

@vars: x, dx, w, prim, sw, h, b, sh, sb

Kinetic.ControlVolume1D3F — Type

1D control volume with 3 distribution functions

ControlVolume1D3F(X, DX, W, PRIM, H0, H1, H2, E0, B0, L)

@vars: x, dx, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

Kinetic.ControlVolume1D4F — Type

1D control volume with 4 distribution functions

ControlVolume1D4F(X, DX, W, PRIM, H0, H1, H2, H3, E0, B0, L)

@vars: x, dx, w, prim, sw, h0, h1, h2, h3, sh0, sh1, sh2, sh3, E, B, ϕ, ψ, lorenz

Kinetic.ControlVolume2D — Type

2D control volume with no distribution function

ControlVolume2D(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw

Kinetic.ControlVolume2D1F — Type

2D control volume with 1 distribution function

ControlVolume2D1F(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray, F::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw, f, sf

Kinetic.ControlVolume2D2F — Type

2D control volume with 2 distribution functions

ControlVolume2D2F(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray, H::AbstractArray, B::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw, h, b, sh, sb

Kinetic.ControlVolume2D3F — Type

2D control volume with 3 distribution functions

ControlVolume2D3F(X, DX, Y, DY, W, PRIM, H0, H1, H2, E0, B0, L)

@vars: x, y, dx, dy, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz