Data Structure

KitBase.ControlVolume1D — Type

1D control volume with no distribution function

ControlVolume1D(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray)

@vars: x, dx, w, prim, sw

KitBase.ControlVolume1D1F — Type

1D control volume with 1 distribution function

ControlVolume1D1F(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray, F::AbstractArray)

@vars: x, dx, w, prim, sw, f, sf

KitBase.ControlVolume1D2F — Type

1D control volume with 2 distribution functions

ControlVolume1D2F(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray, H::AbstractArray, B::AbstractArray)

@vars: x, dx, w, prim, sw, h, b, sh, sb

KitBase.ControlVolume1D3F — Type

1D control volume with 3 distribution functions

ControlVolume1D3F(X, DX, W, PRIM, H0, H1, H2, E0, B0, L)

@vars: x, dx, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

KitBase.ControlVolume1D4F — Type

1D control volume with 4 distribution functions

ControlVolume1D4F(X, DX, W, PRIM, H0, H1, H2, H3, E0, B0, L)

@vars: x, dx, w, prim, sw, h0, h1, h2, h3, sh0, sh1, sh2, sh3, E, B, ϕ, ψ, lorenz

KitBase.ControlVolume2D — Type

2D control volume with no distribution function

ControlVolume2D(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw

KitBase.ControlVolume2D1F — Type

2D control volume with 1 distribution function

ControlVolume2D1F(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray, F::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw, f, sf

KitBase.ControlVolume2D2F — Type

2D control volume with 2 distribution functions

ControlVolume2D2F(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray, H::AbstractArray, B::AbstractArray)

@vars: x, y, dx, dy, w, prim, sw, h, b, sh, sb

KitBase.ControlVolume2D3F — Type

2D control volume with 3 distribution functions

ControlVolume2D3F(X, DX, Y, DY, W, PRIM, H0, H1, H2, E0, B0, L)

@vars: x, y, dx, dy, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

KitBase.Interface1D — Type

1D cell interface with no distribution function

@vars: fw

Interface1D(w::AbstractArray)

KitBase.Interface1D1F — Type

1D cell interface with 1 distribution function

@vars: fw, ff

Interface1D1F(w::AbstractArray, f::AbstractArray)

KitBase.Interface1D2F — Type

1D cell interface with 2 distribution functions

@vars: fw, fh, fb

Interface1D2F(w::AbstractArray, f::AbstractArray)

KitBase.Interface1D3F — Type

1D cell interface with 3 distribution functions

@vars: fw, fh0, fh1, fh2, femL, femR,

deterministic: Interface1D3F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,1})

stochastic: Interface1D3F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,2})

KitBase.Interface1D4F — Type

1D cell interface with 4 distribution functions

@vars: fw, fh0, fh1, fh2, fh3, femL, femR,

deterministic: Interface1D4F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,1})

stochastic: Interface1D4F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,2})

KitBase.Interface2D — Type

2D cell interface with no distribution function

@vars: len, n, fw

Interface2D(L::Real, C::Real, S::Real, w::AbstractArray)

KitBase.Interface2D1F — Type

2D cell interface with 1 distribution function

@vars: len, n, fw, ff

Interface2D1F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

KitBase.Interface2D2F — Type

2D cell interface with 2 distribution functions

@vars: len, n, fw, fh, fb

Interface2D2F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

KitBase.Gas — Type

Gas property

@consts: Kn, Ma, Pr, K, γ, ω, αᵣ, ωᵣ, μᵣ

KitBase.IB — Type

Initial & boundary condition with no distribution function

@consts: wL, primL, bcL, wR, primR, bcR, bcU, bcD

KitBase.IB1F — Type

Initial & boundary condition with 1 distribution function

@consts: wL, primL, fL, bcL, wR, primR, fR, bcR, bcU, bcD

KitBase.IB2F — Type

Initial & boundary condition with 2 distribution functions

@consts: wL, primL, hL, bL, bcL, wR, primR, hR, bR, bcR, bcU, bcD

KitBase.IB3F — Type

Initial & boundary condition with 3 distribution functions

@consts: wL, primL, h0L, h1L, h2L, bcL, EL, BL, lorenzL, wR, primR, h0R, h1R, h2R, bcR, ER, BR, lorenzR, bcU, bcD

KitBase.IB4F — Type

Initial & boundary condition with 4 distribution functions

@consts: wL, primL, h0L, h1L, h2L, h3L, bcL, EL, BL, lorenzL, wR, primR, h0R, h1R, h2R, h3R, bcR, ER, BR, lorenzR, bcU, bcD

KitBase.Mixture — Type

Multi-component gas property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne

KitBase.Plasma1D — Type

1D plasma property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne, lD, rL, sol, χ, ν, Ap, An, D

KitBase.Plasma2D — Type

2D plasma property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne, lD, rL, sol, χ, ν, A1p, A1n, A2p, A2n, D1, D2

KitBase.Setup — Type

Computational setup

@consts: case, space, flux, collision, nSpecies, interpOrder, limiter, cfl, maxTime