General

Kinetic is organized with the data structures and methods of both generality and convenience. While the low-level methods can be applied to multi-dimensional arrays directly, we provide a set of domain-specific structs that handles multiple dispatch in an elegant way.

The struct for setting up the solver is

KitBase.SolverSet — Type

Structure of solver setup

The inner structs of AbstractSolverSet are

KitBase.ControlVolume1D — Type

ControlVolume1D(X::T1, DX::T1, W::T2, PRIM::T2) where {T1<:Real,T2<:AbstractArray}

1D control volume with no distribution function

@vars: x, dx, w, prim, sw

KitBase.ControlVolume1D1F — Type

ControlVolume1D1F(
    X,
    DX,
    W::T1,
    PRIM::T1,
    F::T2,
) where {T1<:AbstractArray,T2<:AbstractArray}

1D control volume with 1 distribution function

@vars: x, dx, w, prim, sw, f, sf

KitBase.ControlVolume1D2F — Type

ControlVolume1D2F(
    X,
    DX,
    W::T1,
    PRIM::T1,
    H::T2,
    B::T2,
) where {T1<:AbstractArray,T2<:AbstractArray}

1D control volume with 2 distribution functions

@vars: x, dx, w, prim, sw, h, b, sh, sb

KitBase.ControlVolume1D3F — Type

ControlVolume1D3F(
    X::Real,
    DX::Real,
    W::AbstractArray{<:Real,2},
    PRIM::AbstractArray{<:Real,2},
    H0::AbstractArray{<:AbstractFloat,3},
    H1::AbstractArray{<:AbstractFloat,3},
    H2::AbstractArray{<:AbstractFloat,3},
    E0::AbstractArray{<:AbstractFloat,1},
    B0::AbstractArray{<:AbstractFloat,1},
    L::AbstractArray{<:AbstractFloat,2},
)

ControlVolume1D3F(
    X::Real,
    DX::Real,
    W::AbstractArray{<:Real,1},
    PRIM::AbstractArray{<:Real,1},
    H0::AbstractArray{<:AbstractFloat,1},
    H1::AbstractArray{<:AbstractFloat,1},
    H2::AbstractArray{<:AbstractFloat,1},
)

ControlVolume1D3F(
    X::Real,
    DX::Real,
    W::AbstractArray{<:Real,3},
    PRIM::AbstractArray{<:Real,3},
    H0::AbstractArray{<:AbstractFloat,4},
    H1::AbstractArray{<:AbstractFloat,4},
    H2::AbstractArray{<:AbstractFloat,4},
    E0::AbstractArray{<:AbstractFloat,2},
    B0::AbstractArray{<:AbstractFloat,2},
    L::AbstractArray{<:AbstractFloat,3},
)

1D control volume with 3 distribution functions

@vars: x, dx, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

KitBase.ControlVolume1D4F — Type

ControlVolume1D4F(
    X::Real,
    DX::Real,
    W::AbstractArray{<:Real,2},
    PRIM::AbstractArray{<:Real,2},
    H0::AbstractArray{<:AbstractFloat,2},
    H1::AbstractArray{Float64,2},
    H2::AbstractArray{Float64,2},
    H3::AbstractArray{Float64,2},
    E0::AbstractArray{Float64,1},
    B0::AbstractArray{Float64,1},
    L::AbstractArray{Float64,2},
)

ControlVolume1D4F(
    X::Real,
    DX::Real,
    W::AbstractArray{<:Real,3},
    PRIM::AbstractArray{<:Real,3},
    H0::AbstractArray{<:AbstractFloat,3},
    H1::AbstractArray{Float64,3},
    H2::AbstractArray{Float64,3},
    H3::AbstractArray{Float64,3},
    E0::AbstractArray{Float64,2},
    B0::AbstractArray{Float64,2},
    L::AbstractArray{Float64,3},
)

1D control volume with 4 distribution functions

@vars: x, dx, w, prim, sw, h0, h1, h2, h3, sh0, sh1, sh2, sh3, E, B, ϕ, ψ, lorenz

KitBase.ControlVolume2D — Type

ControlVolume2D(
    X::Real,
    DX::Real,
    Y::Real,
    DY::Real,
    W::AbstractArray,
    PRIM::AbstractArray,
)

2D control volume with no distribution function

@vars: x, y, dx, dy, w, prim, sw

KitBase.ControlVolume2D1F — Type

ControlVolume2D1F(
    X::Real,
    DX::Real,
    Y::Real,
    DY::Real,
    W::AbstractArray,
    PRIM::AbstractArray,
    F::AbstractArray,
)

2D control volume with 1 distribution function

@vars: x, y, dx, dy, w, prim, sw, f, sf

KitBase.ControlVolume2D2F — Type

ControlVolume2D2F(
    X::Real,
    DX::Real,
    Y::Real,
    DY::Real,
    W::AbstractArray,
    PRIM::AbstractArray,
    H::AbstractArray,
    B::AbstractArray,
)

2D control volume with 2 distribution functions

@vars: x, y, dx, dy, w, prim, sw, h, b, sh, sb

KitBase.ControlVolume2D3F — Type

ControlVolume2D3F(
    X::Real,
    DX::Real,
    Y::Real,
    DY::Real,
    W::AbstractArray,
    PRIM::AbstractArray,
    H0::AbstractArray,
    H1::AbstractArray,
    H2::AbstractArray,
    E0::AbstractArray,
    B0::AbstractArray,
    L::AbstractArray,
)

ControlVolume2D3F(
    X::Real,
    DX::Real,
    Y::Real,
    DY::Real,
    W::AbstractArray{<:Real,1},
    PRIM::AbstractArray{<:Real,1},
    H0::AbstractArray{<:AbstractFloat,2},
    H1::AbstractArray{<:AbstractFloat,2},
    H2::AbstractArray{<:AbstractFloat,2},
)

2D control volume with 3 distribution functions

@vars: x, y, dx, dy, w, prim, sw, h0, h1, h2, sh0, sh1, sh2, E, B, ϕ, ψ, lorenz

KitBase.ControlVolumeParticle1D — Type

ControlVolume1D(X::Real, DX::Real, W::AbstractArray, PRIM::AbstractArray)

1D control volume in correspondence with particle simulation

@vars: x, dx, w, prim, sw, wg, τ, np, vrmax, remainder

KitBase.ControlVolumeParticle2D — Type

ControlVolume2D(X::Real, DX::Real, Y::Real, DY::Real, W::AbstractArray, PRIM::AbstractArray)

2D control volume in correspondence with particle simulation

@vars: x, y, dx, dy, w, prim, sw, wg, τ, np, vrmax, remainder

KitBase.DiatomicGas — Type

Diatomic gas property

@vars: Kn, Ma, Pr, K, γ, ω, αᵣ, ωᵣ, μᵣ, m, np

KitBase.Gas — Type

Particle property

@vars: Kn, Ma, Pr, K, γ, ω, αᵣ, ωᵣ, μᵣ, m, np

KitBase.IB — Type

Initial & boundary condition with no distribution function

@vars: wL, primL, bcL, wR, primR, bcR, bcU, bcD

KitBase.IB1F — Type

Initial & boundary condition with 1 distribution function

@vars: wL, primL, fL, bcL, wR, primR, fR, bcR, bcU, bcD

KitBase.IB2F — Type

Initial & boundary condition with 2 distribution functions

@vars: wL, primL, hL, bL, bcL, wR, primR, hR, bR, bcR, bcU, bcD

KitBase.IB3F — Type

Initial & boundary condition with 3 distribution functions

@vars: wL, primL, h0L, h1L, h2L, bcL, EL, BL, lorenzL, wR, primR, h0R, h1R, h2R, bcR, ER, BR, lorenzR, bcU, bcD

KitBase.IB4F — Type

Initial & boundary condition with 4 distribution functions

@vars: wL, primL, h0L, h1L, h2L, h3L, bcL, EL, BL, lorenzL, wR, primR, h0R, h1R, h2R, h3R, bcR, ER, BR, lorenzR, bcU, bcD

KitBase.Interface1D — Type

Interface1D(w::AbstractArray)

1D cell interface with no distribution function

@vars: fw

KitBase.Interface1D1F — Type

Interface1D1F(w::AbstractArray, f::AbstractArray)

1D cell interface with 1 distribution function

@vars: fw, ff

KitBase.Interface1D2F — Type

Interface1D2F(w::AbstractArray, f::AbstractArray)

1D cell interface with 2 distribution functions

@vars: fw, fh, fb

KitBase.Interface1D3F — Type

Interface1D3F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,1})
Interface1D3F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,2})
Interface1D3F(w::AbstractArray, f::AbstractArray)

1D cell interface with 3 distribution functions

@vars: fw, fh0, fh1, fh2, femL, femR,

KitBase.Interface1D4F — Type

Interface1D4F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,1})
Interface1D4F(w::AbstractArray, f::AbstractArray, E::AbstractArray{<:Real,2})

1D cell interface with 4 distribution functions

@vars: fw, fh0, fh1, fh2, fh3, femL, femR

KitBase.Interface2D — Type

Interface2D(L::Real, C::Real, S::Real, w::AbstractArray)

2D cell interface with no distribution function

@vars: len, n, fw

KitBase.Interface2D1F — Type

Interface2D1F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

2D cell interface with 1 distribution function

@vars: len, n, fw, ff

KitBase.Interface2D2F — Type

Interface2D2F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

2D cell interface with 2 distribution functions

@vars: len, n, fw, fh, fb

KitBase.Interface2D3F — Type

Interface2D3F(
    L::Real,
    C::Real,
    S::Real,
    w::AbstractArray,
    f::AbstractArray,
    E::AbstractArray{<:Real,1},
)
Interface2D3F(
    L::Real,
    C::Real,
    S::Real,
    w::AbstractArray,
    f::AbstractArray,
    E::AbstractArray{<:Real,2},
)
Interface2D3F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)

2D cell interface with 3 distribution functions

@vars: len, n, fw, fh, fb

KitBase.Mixture — Type

Multi-component gas property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne

KitBase.Particle — Type

Particle(M, X, V, E, IDX, FLAG = zeros(eltype(IDX), axes(IDX, 1)), T = zero(M))

Struct of arrays for particle simulation

@vars: m, x, v, e, idx, flag, tc

KitBase.Particle1D — Type

Particle1D(M::AbstractFloat, X::Real, V::AbstractArray, IDX::Integer, T::Real)

1D particle

@vars: m, x, v, e, idx, tc

KitBase.Particle2D — Type

Particle2D(M::AbstractFloat, X::Real, V::AbstractArray, IDX::Integer, T::Real)

2D particle

@vars: m, x, v, idx, tb

KitBase.Plasma1D — Type

1D plasma property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne, lD, rL, sol, χ, ν, Ap, An, D

KitBase.Plasma2D — Type

2D plasma property

@consts: Kn, Ma, Pr, K, γ, mi, ni, me, ne, lD, rL, sol, χ, ν, A1p, A1n, A2p, A2n, D1, D2

KitBase.Setup — Type

Computational setup

@consts: case, space, flux, collision, nSpecies, interpOrder, limiter, cfl, maxTime