Finite volume data

In the finite volume method, the data is stored separately throughout the cells. Therefore, we provide AbstractControlVolume and AbstractInterface structs for processing in-cell and edge information, which are used as arrays of structs (AoS) in numerical simulations. Considering one-dimensional physical space $x$, we provide the following control volume structs. The structs differs from the number of particle distribution functions.

KitBase.ControlVolume1D — Type

mutable struct ControlVolume1D{F,A,B} <: AbstractControlVolume1D
    x::F
    dx::F

    w::A
    prim::B
    sw::A
end

1D control volume with no distribution function

KitBase.ControlVolume1D1F — Type

mutable struct ControlVolume1D1F{F,A,B} <: AbstractControlVolume1D
    x::F
    dx::F

    w::A
    prim::A
    sw::A

    f::B
    sf::B
end

1D control volume with 1 distribution function

KitBase.ControlVolume1D2F — Type

mutable struct ControlVolume1D2F{F,A,B} <: AbstractControlVolume1D
    x::F
    dx::F

    w::A
    prim::A
    sw::A

    h::B
    b::B
    sh::B
    sb::B
end

1D control volume with 2 distribution functions

KitBase.ControlVolume1D3F — Type

mutable struct ControlVolume1D3F{F,A,B,C,D,E} <: AbstractControlVolume1D
    x::F
    dx::F

    w::A
    prim::A
    sw::A

    h0::B
    h1::B
    h2::B
    sh0::B
    sh1::B
    sh2::B

    E::C
    B::C
    ϕ::D
    ψ::D
    lorenz::E
end

1D control volume with 3 distribution functions

KitBase.ControlVolume1D4F — Type

mutable struct ControlVolume1D4F{F,A,B,C,D,E} <: AbstractControlVolume1D
    x::F
    dx::F

    w::A
    prim::A
    sw::A

    h0::B
    h1::B
    h2::B
    h3::B
    sh0::B
    sh1::B
    sh2::B
    sh3::B

    E::C
    B::C
    ϕ::D
    ψ::D
    lorenz::E
end

1D control volume with 4 distribution functions

Within each cell, different numbers of particle distribution function can be tracked. The interface data is stored correspondingly.

KitBase.Interface1D — Type

mutable struct Interface1D{A<:Union{Real,AbstractArray}} <: AbstractInterface1D
    fw::A
end

1D cell interface with no distribution function Note that deepcopy constructor is needed to work with StructArrays

KitBase.Interface1D1F — Type

mutable struct Interface1D1F{A,B} <: AbstractInterface1D
    fw::A
    ff::B
end

1D cell interface with 1 distribution function Note that deepcopy constructor is needed to work with StructArrays

KitBase.Interface1D2F — Type

mutable struct Interface1D2F{A,B} <: AbstractInterface1D
    fw::A
    fh::B
    fb::B

    function Interface1D2F(w::AbstractArray, f::AbstractArray)
        fw = zero(w)
        fh = zero(f)
        fb = zero(f)

        new{typeof(fw),typeof(fh)}(fw, fh, fb)
    end
end

1D cell interface with 2 distribution functions

KitBase.Interface1D3F — Type

mutable struct Interface1D3F{A,B,C} <: AbstractInterface1D
    fw::A
    fh0::B
    fh1::B
    fh2::B
    femL::C
    femR::C
end

1D cell interface with 3 distribution functions

KitBase.Interface1D4F — Type

mutable struct Interface1D4F{A,B,C} <: AbstractInterface1D
    fw::A
    fh0::B
    fh1::B
    fh2::B
    fh3::B
    femL::C
    femR::C
end

1D cell interface with 4 distribution functions

The 2D control volume structs are implemented as well.

KitBase.ControlVolume2D — Type

mutable struct ControlVolume2D{F,A,B} <: AbstractControlVolume2D
    x::F
    y::F
    dx::F
    dy::F

    w::A
    prim::A
    sw::B
end

2D control volume with no distribution function

KitBase.ControlVolume2D1F — Type

mutable struct ControlVolume2D1F{F,A,B,C,D} <: AbstractControlVolume2D
    x::F
    y::F
    dx::F
    dy::F

    w::A
    prim::A
    sw::B

    f::C
    sf::D
end

2D control volume with 1 distribution function

KitBase.ControlVolume2D2F — Type

mutable struct ControlVolume2D2F{F,A,B,C,D} <: AbstractControlVolume2D
    x::F
    y::F
    dx::F
    dy::F

    w::A
    prim::A
    sw::B

    h::C
    b::C
    sh::D
    sb::D
end

2D control volume with 2 distribution functions

KitBase.ControlVolume2D3F — Type

mutable struct ControlVolume2D3F{T1,T2,T3,T4,T5,T6,T7,T8} <: AbstractControlVolume2D
    x::T1
    y::T1
    dx::T1
    dy::T1

    w::T2
    prim::T2
    sw::T3

    h0::T4
    h1::T4
    h2::T4
    sh0::T5
    sh1::T5
    sh2::T5

    E::T6
    B::T6
    ϕ::T7
    ψ::T7
    lorenz::T8
end

2D control volume with 3 distribution functions

The numerical fluxes are evaluated through AbstractInterface structs.

KitBase.Interface2D — Type

mutable struct Interface2D{A,B,C} <: AbstractInterface2D
    len::A
    n::B
    fw::C

    function Interface2D(L::Real, C::Real, S::Real, w::AbstractArray)
        len = L
        n = [C, S]

        fw = zero(w)

        new{typeof(len),typeof(n),typeof(fw)}(len, n, fw)
    end
end

2D cell interface with no distribution function

KitBase.Interface2D1F — Type

mutable struct Interface2D1F{A,B,C,D} <: AbstractInterface2D
    len::A
    n::B
    fw::C
    ff::D

    function Interface2D1F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)
        len = L
        n = [C, S]

        fw = zero(w)
        ff = zero(f)

        new{typeof(len),typeof(n),typeof(fw),typeof(ff)}(len, n, fw, ff)
    end
end

2D cell interface with 1 distribution function

KitBase.Interface2D2F — Type

mutable struct Interface2D2F{A,B,C,D} <: AbstractInterface2D
    len::A
    n::B
    fw::C
    fh::D
    fb::D

    function Interface2D2F(L::Real, C::Real, S::Real, w::AbstractArray, f::AbstractArray)
        len = L
        n = [C, S]

        fw = zeros(eltype(w), axes(w))
        fh = zeros(eltype(f), axes(f))
        fb = zeros(eltype(f), axes(f))

        new{typeof(len),typeof(n),typeof(fw),typeof(fh)}(len, n, fw, fh, fb)
    end
end

2D cell interface with 2 distribution functions