Initial and boundary conditions

A struct ib <: AbstractCondition defines the initial and boundary conditions of a simulation. It contains the values of conservative and primitive variables, and particle distribution functions at left and right (up and down) domain for both initial and boundary conditions. It is set this way to easily deal with discontinuous initial conditions.

KitBase.IB — Type

mutable struct IB{TF, T, NT} <: KitBase.AbstractCondition

Initial & boundary conditions with no distribution function

Fields

fw

bc

p

KitBase.IB1F — Type

mutable struct IB1F{TF1, TF2, T, NT} <: KitBase.AbstractCondition

Initial & boundary conditions with 1 distribution function

Fields

fw

ff

bc

p

KitBase.IB2F — Type

mutable struct IB2F{TF1, TF2, T, NT} <: KitBase.AbstractCondition

Initial & boundary conditions with 2 distribution functions

Fields

fw

ff

bc

p

KitBase.IB3F — Type

mutable struct IB3F{TF1, TF2, TF3, TF4, TF5, T, NT} <: KitBase.AbstractCondition

Initial & boundary conditions with 3 distribution functions

Fields

fw

ff

fE

fB

fL

bc

p

KitBase.IB4F — Type

mutable struct IB4F{TF1, TF2, TF3, TF4, TF5, T, NT} <: KitBase.AbstractCondition

Initial & boundary conditions with 4 distribution functions

Fields

fw

ff

fE

fB

fL

bc

p