Particle velocity space

A struct vSpace <: AbstractSetup defines the particle velocity setup of a simulation. Structs for 1-3 dimensional particle velocity space are built.

KitBase.VSpace1D — Type

struct VSpace1D{TR<:Real,TI<:Integer,TA<:AbstractArray,TB<:AbstractArray{<:Real,1}} <: AbstractVelocitySpace
    u0::TR
    u1::TR
    nu::TI
    u::TA
    du::TA
    weights::TB
end

1D velocity space

KitBase.VSpace2D — Type

struct VSpace2D{TR<:Real,TI<:Integer,TA<:AbstractArray{<:Real,2}} <: AbstractVelocitySpace
    u0::TR
    u1::TR
    nu::TI
    v0::TR
    v1::TR
    nv::TI
    u::TA
    v::TA
    du::TA
    dv::TA
    weights::TA
end

2D velocity space

KitBase.VSpace3D — Type

struct VSpace3D{TR<:Real,TI<:Integer,TA<:AbstractArray{<:Real,3}} <: AbstractVelocitySpace
    u0::TR
    u1::TR
    nu::TI
    v0::TR
    v1::TR
    nv::TI
    w0::TR
    w1::TR
    nw::TI
    u::TA
    v::TA
    w::TA
    du::TA
    dv::TA
    dw::TA
    weights::TA
end

3D velocity space

It contains

u0 (v0, w0): location of starting point

u1 (v1, w1): location of ending point

nu (nv, nw): number of cells in one direction

u (v, w): locations of middle points of all cells

du (dv, dw): intervals of all cell points

weights: quadrature weights for numerical integral

Note that the one-dimensional velocity space can be used to handle 1-3 dimensional unstructured topology as well. In addition, velocity space structs for multi-component substance are implemented.

KitBase.MVSpace1D — Type

struct MVSpace1D{TR<:AbstractArray{<:Real,1},TI<:Integer,TA<:AbstractArray{<:Real,2}} <: AbstractVelocitySpace
    u0::TR
    u1::TR
    nu::TI
    u::TA
    du::TA
    weights::TA
end

1D multi-component velocity space

KitBase.MVSpace2D — Type

struct MVSpace2D{TR<:AbstractArray{<:Real,1},TI<:Integer,TA<:AbstractArray{Float64,3}} <: AbstractVelocitySpace
    u0::TR
    u1::TR
    nu::TI
    v0::TR
    v1::TR
    nv::TI
    u::TA
    v::TA
    du::TA
    dv::TA
    weights::TA
end

2D multi-component velocity space

KitBase.MVSpace3D — Type

struct MVSpace3D{TR<:AbstractArray{<:Real,1},TI<:Integer,TA<:AbstractArray{<:Real,4}} <: AbstractVelocitySpace
    u0::TR
    u1::TR
    nu::TI
    v0::TR
    v1::TR
    nv::TI
    w0::TR
    w1::TR
    nw::TI
    u::TA
    v::TA
    w::TA
    du::TA
    dv::TA
    dw::TA
    weights::TA
end

3D multi-component velocity space

For the simulation cases where no phase-space evolution is involved, vSpace can be set as nothing directly.

vSpace = nothing